% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartMapConstructor.R
\name{add_ssimChart}
\alias{add_ssimChart}
\title{Customize the chart}
\usage{
add_ssimChart(c, object, objectname)
}
\arguments{
\item{object}{An object to add to the chart}

\item{plot}{The chart object to add \code{object} to}

\item{object_name}{The name of the object to add}
}
\value{
A modified chart object
}
\description{
This generic allows you to add your own methods for adding custom objects to
a chart with \link{+.Chart}.
}
\keyword{internal}
