% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{addModule}
\alias{addModule}
\title{Add module}
\usage{
addModule(filename, session = NULL)
}
\arguments{
\item{filename}{character string or vector of these. The path to an .ssimpkg 
file on disk, or a vector of filepaths.}

\item{session}{\code{\link{Session}} object}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
Please use \code{\link{addPackage}} instead.
}
\keyword{internal}
