% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{addModules<-}
\alias{addModules<-}
\title{Add modules}
\usage{
addModules(x) <- value
}
\arguments{
\item{x}{A SyncroSim \code{\link{Session}} object.}

\item{value}{The path to an .ssimpkg file on disk, or a vector of filepaths}
}
\description{
Add module or modules to this version of SyncroSim
}

