% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addon.R
\name{addon}
\alias{addon}
\alias{addon,character-method}
\alias{addon,missingOrNULL-method}
\alias{addon,Session-method}
\alias{addon,SsimObject-method}
\title{addon(s) of a SsimLibrary or Session}
\usage{
addon(ssimObject)

\S4method{addon}{character}(ssimObject)

\S4method{addon}{missingOrNULL}(ssimObject)

\S4method{addon}{Session}(ssimObject)

\S4method{addon}{SsimObject}(ssimObject)
}
\arguments{
\item{ssimObject}{An object of class SsimLibrary/Project/Scenario or Session.}
}
\value{
A dataframe listing the addon(s) in use by the library to which the object
belongs.
}
\description{
Some Packages are Add-On Packages. These Packages can only be used to extend
existing Base Packages; as a result they cannot be used to create new Libraries.
For example, stsimsf is an Add-On Package for stsim which provides optional
additional functionality for the base ST-Sim model.
More information on add-ons can be found in the
\href{http://docs.syncrosim.com/how_to_guides/package_addon.html}{syncrosim documentation}.
}
\details{
The addon(s) of a \code{\link{SsimLibrary}} or \code{\link{Session}}.
}
\examples{
\donttest{
addPackage("stsim")
temp_dir <- tempdir()
mySession <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"),
                         session = mySession)

addon(myLibrary)
}

}
