% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPackageFile.R
\docType{methods}
\name{addPackageFile}
\alias{addPackageFile}
\alias{addPackageFile,character-method}
\title{Adds a package to SyncroSim}
\usage{
addPackageFile(filename = NULL, session = NULL)

\S4method{addPackageFile}{character}(filename = NULL, session = NULL)
}
\arguments{
\item{filename}{Character string.  The path to a SyncroSim package file.}

\item{session}{Session.}
}
\description{
Adds a package to SyncroSim
}
