% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRow.R
\name{addRow<-}
\alias{addRow<-}
\title{Add a row to a datasheet.}
\usage{
addRow(x) <- value
}
\arguments{
\item{x}{A dataframe.}

\item{value}{A dataframe. Columns in should be a subset of columns in x.}
}
\value{
A dataframe with a new row.
}
\description{
Adds a row to a dataframe.
Missing values are filled if possible using factor levels.
}

