% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRows.R
\name{addRows<-}
\alias{addRows<-}
\title{Add rows to a datasheet.}
\usage{
addRows(x) <- value
}
\arguments{
\item{x}{A dataframe.}

\item{value}{A dataframe. Columns in value should be a subset of columns in x.}
}
\value{
A dataframe with new rows.
}
\description{
Adds rows to a dataframe.
Preserves the types and factor levels of x.
Fills missing values if possible using factor levels.
}

