% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allCombos.R
\name{allCombos}
\alias{allCombos}
\title{Get all combinations of factor levels}
\usage{
allCombos(x)
}
\arguments{
\item{x}{An empty data frame with factors.}
}
\value{
A data frame containing all possible combinations of factor levels.
}
\description{
Given an empty data frame, get all possible combinations of factor levels
}

