% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogentags.R
\name{autogentags}
\alias{autogentags}
\alias{autogentags,character-method}
\alias{autogentags,Scenario-method}
\title{Auto Generation Tags for a Scenario.}
\usage{
autogentags(ssimObject)

\S4method{autogentags}{character}(ssimObject)

\S4method{autogentags}{Scenario}(ssimObject)
}
\arguments{
\item{ssimObject}{Scenario.}
}
\value{
Returns the Auto Generation Tags.
}
\description{
Retrieves the Auto Generation Tags for a Scenario.
}
