% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{basePackage}
\alias{basePackage}
\title{Installed base packages}
\usage{
basePackage(ssimObject = NULL)
}
\arguments{
\item{ssimObject}{\code{\link{Session}} or \code{\link{SsimLibrary}} object}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Please use \code{\link{package}} instead.
}
\keyword{internal}
