% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindToEnv.R
\name{bindToEnv}
\alias{bindToEnv}
\title{Copy arguments into env and re-bind any function's lexical scope to bindTargetEnv .}
\usage{
bindToEnv(bindTargetEnv = parent.frame(), objNames, doNotRebind = c())
}
\arguments{
\item{bindTargetEnv}{environment to bind to}

\item{objNames}{additional names to lookup in parent environment and bind}

\item{names}{of functions to NOT rebind the lexical environments of}
}
\description{
See http://winvector.github.io/Parallel/PExample.html for example use.
}
\details{
Used to send data along with a function in situations such as parallel execution
(when the global environment would not be available).  Typically called within
a function that constructs the worker function to pass to the parallel processes
(so we have a nice lexical closure to work with).
}

