% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint.R
\docType{class}
\name{Breakpoint-class}
\alias{Breakpoint}
\alias{Breakpoint-class}
\title{Breakpoint class}
\description{
Breakpoint class
}
\section{Slots}{

\describe{
\item{\code{arguments}}{Timesteps or iterations e.g. "1,2"}

\item{\code{breakpointName}}{Breakpoint name}

\item{\code{name}}{Name}

\item{\code{transformerName}}{'stsim:core-transformer' or?}

\item{\code{callback}}{The function to apply. See STSimBreakpointsTutorial.R for details.}
}}

