% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{breakpointInfo<-}
\alias{breakpointInfo<-}
\title{Set breakpointInfo of a Scenario.}
\usage{
breakpointInfo(x) <- value
}
\arguments{
\item{x}{A SyncroSim Scenario}

\item{value}{A list of breakpoint information. See details.}
}
\value{
An SyncroSim Scenario object containing breakpoints
}
\description{
Set breakpointInfo for a SyncroSim Scenario.
}
\details{
breakpointInfo is a list of the following form:
list(breakpoints=myVectorOfBreakpoints,breakpointFunction = myBreakpointFunction)

myVectorOfBreakpoints is a vector of integers representing timesteps. SyncroSim will apply myBreakpointFunction at each timestep.
myBreakpointFunction is the function to apply. See STSimBreakpointsTutorial.R for details.
}

