% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{breakpointInfo}
\alias{breakpointInfo}
\title{The breakpointInfo of a Scenario}
\usage{
breakpointInfo(x)
}
\arguments{
\item{x}{A Scenario object.}
}
\value{
A list of breakpointInfo.
}
\description{
The breakpointInfo of a Scenario
}
\details{
breakpointInfo is a list of the following form:
list(breakpoints=myVectorOfBreakpoints,breakpointFunction = myBreakpointFunction)

myVectorOfBreakpoints is a vector of integers representing timesteps. SyncroSim will apply myBreakpointFunction at each timestep.
myBreakpointFunction is the function to apply. See STSimBreakpointsTutorial.R for details.
}

