% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{breakpoints}
\alias{breakpoints}
\title{The breakpoints of a Scenario}
\usage{
breakpoints(x)
}
\arguments{
\item{x}{A Scenario object.}
}
\value{
A list of Breakpoint objects.
}
\description{
The breakpoints of a Scenario
}

