% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointSession.R
\docType{class}
\name{BreakpointSession-class}
\alias{BreakpointSession-class}
\alias{BreakpointSession}
\title{BreakpointSession class}
\description{
BreakpointSession class
}
\section{Slots}{

\describe{
\item{\code{scenario}}{A SyncroSim scenario}

\item{\code{connection}}{A socket connection to the SyncroSim server.}

\item{\code{name}}{A name}
}}

