% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartMapConstructor.R
\name{+.Chart}
\alias{+.Chart}
\title{Testing out plotting construction code
Below taken and modified from ggplot plot-construction.R}
\usage{
\method{+}{Chart}(e1, e2)

`+.Chart`(e1, e2)
}
\arguments{
\item{e1}{An object of class \code{\link[=ssimChart]{ssimChart()}}.}

\item{e2}{A chart or map component, as described below.}
}
\description{
Create function for adding together SyncroSim charting and mapping objects
}
