% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartInfo.R
\name{chartInfo}
\alias{chartInfo}
\alias{chartInfo,SsimObject-method}
\title{Retrieves chart variables}
\usage{
chartInfo(ssimObject, chart = NULL)

\S4method{chartInfo}{SsimObject}(ssimObject, chart = NULL)
}
\arguments{
\item{ssimObject}{\code{\link{Project}} or \code{\link{Chart}} object}

\item{chart}{character or integer. Either the name or ID of an existing chart.
If \code{NULL} and a \code{\link{Project}} is provided as the first argument,
then will return the available variables for charting.}
}
\value{
A data.frame of variables, filter columns, and filter values.
}
\description{
Retrieves the available variables for charting, or the variables that are
set for an existing chart.
}
\details{
Example arguments:
\itemize{
\item If ssimObject is SyncroSim Project and chart is \code{NULL}: Returns
a data.frame of available variables for creating a new chart.
\item If ssimObject is SyncroSIm Chart or chart is not \code{NULL}: Returns
a data.frame of variables in use by the specified chart.
}
}
\examples{
\dontrun{
# Create a chart object
myChart <- chart(myProject, chart = "New Chart")

# Retrieve variables that can be used to create new charts
chartInfo(myProject)

# Retrieve variables being used by existing chart
chartInfo(myChart)
}

}
