% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartLine.R
\name{chartLine}
\alias{chartLine}
\alias{chartLine,Chart-method}
\title{Sets the \code{\link{Chart}} type to "Line"}
\usage{
chartLine(
  chart,
  x = "Timesteps",
  y = NULL,
  timesteps = NULL,
  iterationType = "mean",
  iteration = 1
)

\S4method{chartLine}{Chart}(
  chart,
  x = "Timesteps",
  y = NULL,
  timesteps = NULL,
  iterationType = "mean",
  iteration = 1
)
}
\arguments{
\item{x}{character or character vector. X variable(s) to plot (Default is
"Timesteps").}

\item{y}{character or character vector. Y variable(s) to plot.}

\item{timesteps}{integer vector. The range of timesteps to plot against
(Default is the minimum and maximum timesteps defined in the simulation).}

\item{iterationType}{character. How to display multiple iterations in the
chart. Can be "mean" (Default), "single", or "all".}

\item{iteration}{integer. If the \code{iterationType} is set to "single", this argument
determines which iteration to display. Default is 1.}

\item{Chart}{\code{\link{Chart}} object}
}
\value{
A \code{Chart} object representing a SyncroSim chart
}
\description{
Sets the \code{\link{Chart}} type to "Line" and adds the variables to plot
in the line chart.
}
\examples{
\dontrun{
# Create a chart object
myChart <- chart(myProject, chart = "New Chart")

# Set the chart type and data
myChart <- chartLine(myChart, x = "Timesteps", y = c("variable1", "variable2"),
timesteps = c(0,10), iterationType = "single", iteration = 1)

}

}
