% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{conda}
\alias{conda}
\alias{conda,Session-method}
\alias{conda,missingOrNULLOrChar-method}
\alias{conda<-}
\alias{conda<-,character-method}
\alias{conda<-,Session-method}
\title{Conda path of SyncroSim Session}
\usage{
conda(session = NULL)

\S4method{conda}{Session}(session = NULL)

\S4method{conda}{missingOrNULLOrChar}(session = NULL)

conda(session) <- value

\S4method{conda}{character}(session) <- value

\S4method{conda}{Session}(session) <- value
}
\arguments{
\item{session}{\code{\link{Session}} object or character (i.e. filepath to a 
session). If \code{NULL}, \code{session()} will be used}

\item{value}{character. If not \code{NULL} (default), the SyncroSim Session 
will use the given path to the Conda installation.}
}
\value{
A character: The set path to the Conda installation.
}
\description{
Checks or sets the SyncroSim \code{\link{Session}} Conda installation Path.
If no path is specified and a package requires Conda environments, then the
default Conda installation path will be used.
}
\examples{
\donttest{
# Set up a SyncroSim Session
mySession <- session()

# Check the Conda path of a SyncroSim Session
conda(mySession)

# Set the Conda path for a SyncroSim Session
conda(mySession) <- "C:/Users/Admin/miniconda3"
}

}
