% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointSession.R
\docType{methods}
\name{connection}
\alias{connection}
\alias{connection,BreakpointSession-method}
\alias{connection,missingOrNULLOrChar-method}
\title{Get or set a socket connection.}
\usage{
connection(x, ...)

\S4method{connection}{missingOrNULLOrChar}(x = "127.0.0.1", port = 13000)

\S4method{connection}{BreakpointSession}(x)
}
\arguments{
\item{x}{An ipAddress or BreakpointSession object. If NULL a default ip will be used.}

\item{port}{For new connections only - a port number.}
}
\description{
Get or set a socket connection.
}
\section{Methods (by class)}{
\itemize{
\item \code{missingOrNULLOrChar}: Get a new connection.

\item \code{BreakpointSession}: Get the connection of a BreakpointSession.
}}

