% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNestedFolder.R
\name{createNestedFolder}
\alias{createNestedFolder}
\alias{createNestedFolder,character-method}
\alias{createNestedFolder,Project-method}
\title{Create a nested folder within another folder}
\usage{
createNestedFolder(ssimObject, parentFolderId, folderName)

\S4method{createNestedFolder}{character}(ssimObject, parentFolderId, folderName)

\S4method{createNestedFolder}{Project}(ssimObject, parentFolderId, folderName)
}
\arguments{
\item{ssimObject}{\code{\link{Project}} object}

\item{parentFolderId}{integer}

\item{folderName}{character string representing folder name}
}
\value{
A Character String: the nested folder ID value
}
\description{
Create a new nested folder for a SyncroSim \code{\link{Project}}. 
The folder will be created within another Project folder and can be
used to organize Scenarios in the SyncroSim User Interface.
}
\examples{
\donttest{
# Specify file path and name of new SsimLibrary
myLibraryName <- file.path(tempdir(), "testlib")

# Set up a SyncroSim Session and SsimLibrary
mySession <- session()
myLibrary <- ssimLibrary(name = myLibraryName, session = mySession)
myProject <- project(ssimLibrary)

# Create a new folder at the Project level and extract the folder Id
parentFolderIdString <- createProjectFolder(myProject, "NewFolder")
parentFolderId <- as.integer(strsplit(parentFolderIdString, ": ")[[1]][2])

# Create nested folder within project folder
createNestedFolder(myProject, parentFolderId, "NewNestedFolder")
}

}
