% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProjectFolder.R
\name{createProjectFolder}
\alias{createProjectFolder}
\alias{createProjectFolder,character-method}
\alias{createProjectFolder,Project-method}
\title{Create folder at the Project level}
\usage{
createProjectFolder(ssimObject, folderName)

\S4method{createProjectFolder}{character}(ssimObject, folderName)

\S4method{createProjectFolder}{Project}(ssimObject, folderName)
}
\arguments{
\item{ssimObject}{\code{\link{Project}} object}

\item{folderName}{character string representing folder name}
}
\value{
A Character String: the folder ID value
}
\description{
Create a new folder for a SyncroSim \code{\link{Project}}. 
The folder will be created at the SyncroSim Project root and can be
used to organize Scenarios in the SyncroSim User Interface.
}
\examples{
\donttest{
# Specify file path and name of new SsimLibrary
myLibraryName <- file.path(tempdir(), "testlib")

# Set up a SyncroSim Session and SsimLibrary
mySession <- session()
myLibrary <- ssimLibrary(name = myLibraryName, session = mySession)
myProject <- project(ssimLibrary, name = "Definitions")

# Create a new folder at the Project level
createProjectFolder(myProject, "NewFolder")
}

}
