% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{datasheets}
\alias{datasheets}
\title{datasheets}
\usage{
datasheets(x, ...)
}
\arguments{
\item{x}{An SSimLibrary, Project or Scenario object. Or a path to a SyncroSim library on disk.}

\item{project}{Project name or id. Ignored if x is a Project.}

\item{scenario}{Scenario name or id. Ignored if x is a Scenario.}

\item{names}{If TRUE (default) returns dataframe of sheet names. If false returns a named list of dataframes representing each datasheet.}

\item{empty}{If TRUE returns empty template dataframes.}

\item{scope}{"scenario","project", "library", or NULL.}
}
\value{
A dataframe of datasheet names, or list of datasheets represented by dataframes.
}
\description{
Gets datasheets from an SSimLibrary, Project or Scenario.
}
\details{
\itemize{
  \item {If x/project/scenario identify a scenario: }{Returns library, project, and scenario scope datasheets.}
  \item {If x/project/scenario identify a project (but not a scenario): }{Returns library and project scope datasheets.}
  \item {If x/project/scenario identify a library (but not a project or scenario): }{Returns library scope datasheets.}
}
}
\examples{

}

