% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultModel.R
\name{defaultModel<-}
\alias{defaultModel<-}
\title{Set defaultModel of a Session}
\usage{
defaultModel(session) <- value
}
\arguments{
\item{session}{Session.}

\item{value}{character. A SyncroSim model. See model() for options.}
}
\description{
Set defaultModel of a session
}

