% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{definitions}
\alias{definitions}
\title{definitions}
\usage{
datasheets(x, project = NULL, scenario = NULL, names = T, scope = NULL,
  optional = F, empty = F, sheetNames = NULL, lookupsAsFactors = T)
}
\description{
Alias for \code{\link{datasheets}} function
}

