% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint.R
\name{deleteBreakpoint}
\alias{deleteBreakpoint}
\title{Delete a Scenario breakpoint.}
\usage{
deleteBreakpoint(x, transformerName = NULL, breakpointType = NULL)
}
\arguments{
\item{x}{A SyncroSim Scenario}

\item{transformerName}{A Stochastic Time Transformer (e.g. stsim:runtime).  Optional.}

\item{breakpointType}{bi: before iteration; ai: after iteration; bt:before timestep; at: after timestep.  Optional.}
}
\value{
A SyncroSim Scenario with an updated list of breakpoints
}
\description{
This function will delete a Scenario breakpoint.
}
\examples{
myScenario = clearBreakpoint(myScenario)                                    #Deletes all breakpoints
myScenario = clearBreakpoint(myScenario, transformerName="stsim:runtime")   #Deletes breakpoints for stsim:runtime
}
