% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{deleteModule}
\alias{deleteModule}
\title{Delete module or modules}
\usage{
deleteModule(name, session = NULL)
}
\arguments{
\item{name}{A module or vector of modules to remove. \code{\link{modules()}} for options.}

\item{session}{A SyncroSim \code{\link{Session}} object.}
}
\description{
Delete module or modules from this version of SyncroSim.
Note that removing a module can be difficult to undo.
To restore the module the user will need to provide a .ssimpkg file or reinstall SyncroSim.
Thus, \code{deleteModule} requires confirmation from the user.
}

