% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletePackage.R
\name{deletePackage}
\alias{deletePackage}
\alias{deletePackage,ANY,character-method}
\alias{deletePackage,ANY,missingOrNULL-method}
\alias{deletePackage,ANY,Session-method}
\title{Deletes a package}
\usage{
deletePackage(name, session = NULL, force = FALSE)

\S4method{deletePackage}{ANY,character}(name, session = NULL, force = FALSE)

\S4method{deletePackage}{ANY,missingOrNULL}(name, session = NULL, force = FALSE)

\S4method{deletePackage}{ANY,Session}(name, session = NULL, force = FALSE)
}
\arguments{
\item{name}{Character. The name of the package to delete.}

\item{session}{An object of class \code{\link{Session}}.}

\item{force}{Logical. If TRUE, delete without requiring confirmation from 
the user.}
}
\value{
This function invisibly returns `TRUE` upon success (i.e.successful 
deletion) and `FALSE` upon failure.
}
\description{
Deletes a package from your SyncroSim installation.
}
\examples{
\donttest{
temp_dir <- tempdir()
mySession <- session()

deletePackage("stsim", mySession, force = FALSE)
}

}
