% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimLibrary.R
\name{deleteScenarios}
\alias{deleteScenarios}
\title{Delete scenarios}
\usage{
deleteScenarios(x, scenario = NULL, force = FALSE)
}
\arguments{
\item{x}{An SsimLibrary, Project or Scenario.}

\item{scenario}{One or more scenario names or ids.}

\item{force}{If FALSE (default) user is prompted to confirm deletions.}
}
\value{
A list of "saved" or failure messages for each scenario.
}
\description{
Deletes one or more scenarios from a SyncroSim library.
}
\examples{
myLibrary = ssimLibrary()
myScenario = scenario(project(myLibrary))
scenarios(myLibrary,names=T)
deleteScenarios(myLibrary,scenario="Scenario")
scenarios(myLibrary,names=T)
}

