% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{description<-}
\alias{description<-}
\alias{description<-,character-method}
\alias{description<-,SsimObject-method}
\title{Set the description of a SsimLibrary, Project or Scenario.}
\usage{
description(ssimObject) <- value

\S4method{description}{character}(ssimObject) <- value

\S4method{description}{SsimObject}(ssimObject) <- value
}
\arguments{
\item{ssimObject}{Scenario/Project/SsimLibrary.}

\item{value}{The new description.}
}
\value{
The object with updated description.
}
\description{
Set the description of a \code{\link{SsimLibrary}}, \code{\link{Project}} or 
\code{\link{Scenario}}.
}
\examples{
\donttest{
temp_dir <- tempdir()
mySession <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"), session = mySession)
myProject <- project(myLibrary)

description(myProject) <- mydescription
}

}
