% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimLibrary.R
\name{disableAddons<-}
\alias{disableAddons<-}
\title{Disable addons.}
\usage{
disableAddons(x) <- value
}
\arguments{
\item{x=}{A SSimLibrary, Project or Scenario.}
}
\value{
x
}
\description{
Disable addons an SSimLibrary, or Project/Scenario with an associated SSimLibrary.
}
\examples{
myLibrary = ssimLibrary()
enableAddons(myLibrary)=c("stsim-ecological-departure")
addons(myLibrary)
disableAddons(myLibrary)=c("stsim-ecological-departure")
addons(myLibrary)

}

