% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimLibrary.R
\name{enableAddon}
\alias{enableAddon}
\title{Enable addon or addons.}
\usage{
enableAddon(ssimLibrary, name)
}
\arguments{
\item{ssimLibrary}{SsimLibrary}

\item{name}{Character string or vector of these.}
}
\value{
saved or error message for each addon.
}
\description{
Enable addon or addons of an SsimLibrary.
}
\examples{
myLibrary = ssimLibrary()
enableAddon(myLibrary,c("stsim-ecological-departure", "stsim-stock-flow"))
addons(myLibrary)
}

