% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimLibrary.R
\name{enableAddons<-}
\alias{enableAddons<-}
\title{Enable addons.}
\usage{
enableAddons(x) <- value
}
\arguments{
\item{x=}{A SsimLibrary, Project or Scenario.}
}
\value{
x
}
\description{
Enable addons of an SsimLibrary, or Project/Scenario with an associated SsimLibrary.
}
\examples{
myLibrary = ssimLibrary()
enableAddons(myLibrary)=c("stsim-ecological-departure", "stsim-stock-flow")
addons(myLibrary)
}

