% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{envBeginSimulation}
\alias{envBeginSimulation}
\title{Begins a SyncroSim simulation}
\usage{
envBeginSimulation(totalSteps)
}
\arguments{
\item{totalSteps}{integer.  The total number of steps in the simulation}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Please use \code{\link{progressBar}} instead.
}
\keyword{internal}
