% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{envEndSimulation}
\alias{envEndSimulation}
\title{Ends a SyncroSim simulation}
\usage{
envEndSimulation()
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
Please use \code{\link{progressBar}} instead.
}
\keyword{internal}
