% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{envInputFolder}
\alias{envInputFolder}
\title{SyncroSim DataSheet Input Folder}
\usage{
envInputFolder(scenario, datasheetName)
}
\arguments{
\item{scenario}{\code{\link{Scenario}} object. A SyncroSim result Scenario}

\item{datasheetName}{character. The input Datasheet name}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Please use \code{\link{runtimeInputFolder}} instead.
}
\keyword{internal}
