% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{envStepSimulation}
\alias{envStepSimulation}
\title{Steps a SyncroSim simulation}
\usage{
envStepSimulation()
}
\value{
No returned value, used for side effects.
}
\description{
This function is part of a set of functions designed to facilitate the
development of R-based Syncrosim Packages. This function steps a SyncroSim 
simulation, and is to be called between \link{\code{envBeginSimulation}} and 
\link{\code{envEndSimulation}}.
}
\examples{
\donttest{
envStepSimulation()
}

}
