% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{envTempFolder}
\alias{envTempFolder}
\title{SyncroSim Temporary Folder}
\usage{
envTempFolder(folderName)
}
\arguments{
\item{folderName}{character. The folder name}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
Please use \code{\link{runtimeTempFolder}} instead.
}
\keyword{internal}
