% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignoreDependencies.R
\name{ignoreDependencies<-}
\alias{ignoreDependencies<-}
\alias{ignoreDependencies<-,character-method}
\alias{ignoreDependencies<-,Scenario-method}
\title{Ignore Dependencies Datafeeds for a Scenario.}
\usage{
ignoreDependencies(ssimObject) <- value

\S4method{ignoreDependencies}{character}(ssimObject) <- value

\S4method{ignoreDependencies}{Scenario}(ssimObject) <- value
}
\arguments{
\item{ssimObject}{An object of class \code{\link{Scenario}}.}

\item{value}{The name of the data}
}
\value{
Returns the updated ssimObject.
}
\description{
Sets the Ignore Dependencies Datafeed for a \code{\link{Scenario}}.
}
\examples{
\donttest{
temp_dir <- tempdir()
mySession <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"), session = mySession)
myProject <- project(myLibrary)
myScenario <- scenario(myProject)

myScenario <- ignoreDependencies(myScenario, NULL)
}

}
