% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelplotCategorical.R
\name{levelplotCategorical}
\alias{levelplotCategorical}
\title{levelplot of a categorical RasterLayer with a Raster Attribute Table}
\usage{
levelplotCategorical(x, attribute, title = NULL)
}
\arguments{
\item{x}{A Raster object with a raster attribute table set by spatialData() or rasterAttributes().}

\item{attribute}{The attribute to be plotted. This must be a column name in the raster attribute table.}
}
\description{
Plot a RasterLayer with raster attributes set by spatialData() or rasterAttributes().
This is a wrapper around the levelplot() function of the rasterVis package.
}
\examples{

levels(myRaster) #Retrieve the raster attribute table
levelplotCategorical(myRaster,attribute="Name") #plot Name attribute

}

