% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{loadDatasheets}
\alias{loadDatasheets}
\title{Set datasheets}
\usage{
loadDatasheets(x, data, name, project = NULL, scenario = NULL,
  breakpoint = F, printCmd = F)
}
\arguments{
\item{x}{An SsimLibrary, Project or Scenario object. Or the path to a library on disk.}

\item{data}{A dataframe or named list of dataframes to load.}

\item{name}{The sheet name - required if data is a dataframe, ignored otherwise.}

\item{project}{Project name or id.}

\item{scenario}{Scenario name or id.}

\item{breakpoint}{Set to TRUE when modifying datasheets in a breakpoint function.}

\item{printCmd}{Set to TRUE to see the SyncroSim command line arguments. Helpful for debugging.}
}
\value{
A named list of success or failure reports.
}
\description{
Loads datasheets into the SyncroSim library.
}
\examples{

}

