% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{loadSpatialData}
\alias{loadSpatialData}
\title{Set spatial data}
\usage{
loadSpatialData(x, data, metadata, project = NULL, scenario = NULL,
  breakpoint = F, check = T)
}
\arguments{
\item{x}{An SsimLibrary, Project or Scenario object. Or the path to a library on disk.}

\item{data}{A RasterLayer or RasterStack to load.}

\item{metadata}{A dataframe that can be appended to datasheet(x,metadata$SheetName[1]) - see details.}

\item{project}{Project name or id.}

\item{scenario}{Scenario name or id.}

\item{breakpoint}{Set to TRUE when setting spatial data in a breakpoint function.}

\item{check}{Default is TRUE. Set FALSE to speed calculations my assuming metadata is valid.}
}
\value{
A named list of success or failure reports.
}
\description{
Loads spatial data into the SyncroSim library.
}
\details{
Metadata should be a dataframe that can be appended to datasheet(x,metadata$SheetName[1]), containing 1 row for each layer of data.
"SheetName" and at least one "FileName" columns are expected in metadata.
A "RasterLayerName" column in metadata is optional. If present, it should contain the name of each layer in the data raster stack.
If "RasterLayerName" is not included in metadata, names(data) should correspond to the FileNames in metadata.

Note: Spatial data will be appended if non-FileName columns (e.g. Iteration, Timestep, etc) differ between metadata and datasheet(x,metadata$SheetName[1]). Otherwise, the new spatial data will overwrite the old spatial data.
}
\examples{

}

