% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeDependencies.R
\name{mergeDependencies<-}
\alias{mergeDependencies<-}
\alias{mergeDependencies<-,character-method}
\alias{mergeDependencies<-,Scenario-method}
\title{Merge Dependencies for a Scenario.}
\usage{
mergeDependencies(ssimObject) <- value

\S4method{mergeDependencies}{character}(ssimObject) <- value

\S4method{mergeDependencies}{Scenario}(ssimObject) <- value
}
\arguments{
\item{ssimObject}{An object of class \code{\link{Scenario}}.}

\item{value}{Logical. If TRUE the Scenario will be set to merge dependencies at run time.}
}
\value{
Returns the updated ssimObject.
}
\description{
Sets whether or not a \code{\link{Scenario}} is configured to merge 
dependencies at run time.
}
\examples{
\donttest{
temp_dir <- tempdir()
mySession <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"), session = mySession)
myProject <- project(myLibrary, project = "Definitions")
myScenario <- scenario(myProject, scenario = "My Scenario")

mergeDependencies(myScenario) <- TRUE
}

}
