% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{methods}
\name{model-deprecated}
\alias{model-deprecated}
\alias{model,character-method}
\alias{model,missingOrNULL-method}
\alias{model,Session-method}
\alias{model,SsimLibrary-method}
\title{model}
\usage{
\S4method{model}{character}(ssimObject = NULL)

\S4method{model}{missingOrNULL}(ssimObject = NULL)

\S4method{model}{Session}(ssimObject = NULL)

\S4method{model}{SsimLibrary}(ssimObject = NULL)
}
\seealso{
\code{\link{rsyncrosim-deprecated}}
}
\keyword{internal}
