% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{model}
\alias{model}
\title{Installed models}
\usage{
model(ssimObject = NULL)
}
\arguments{
\item{ssimObject}{\code{\link{Session}} or \code{\link{SsimLibrary}} object}
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
Models are now distributed in Packages; 
please use \code{\link{package}} instead.
}
\keyword{internal}
