% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{modelName}
\alias{modelName}
\title{The name of the model associate with a SyncroSim object}
\usage{
modelName(x)
}
\arguments{
\item{x}{An object with an associated model name.}
}
\description{
The name of the model associated with a SSimLibarary, Project or Scenario.
}

