% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimLibrary.R
\name{modelVersion}
\alias{modelVersion}
\title{The version of the primary model associated with a SyncroSim object}
\usage{
modelVersion(x)
}
\arguments{
\item{x}{An object with an associated primary model.}
}
\value{
A model version.
}
\description{
The version of the primary model associated with a SSimLibarary, Project or Scenario.
}

