% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{module}
\alias{module}
\title{Installed modules}
\usage{
module(session = NULL)
}
\arguments{
\item{session}{\code{\link{Session}} object}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
modules are now distributed in Packages;
Please use \code{\link{package}} instead.
}
\keyword{internal}
