% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{modules}
\alias{modules}
\title{Installed modules}
\usage{
modules(session)
}
\arguments{
\item{session}{Session.}
}
\value{
A dataframe of modules
}
\description{
Modules installed with this version of SyncroSim
}

