% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{multiband}
\alias{multiband}
\title{Modify the grouping of spatial layers.}
\usage{
multiband(x, action, grouping = NULL)
}
\arguments{
\item{x}{A SyncroSim results Scenario or list of SyncroSim result Scenarios.}

\item{action}{Options are: apply, remove, rebuild}

\item{grouping}{Only used if action=apply. If NULL use datasheet(myLibrary,name="STime_Options"). Options are: Iteration,Timestep,All}
}
\value{
"saved" or an error message from SyncroSim.
}
\description{
Modify the grouping of spatial output layers in a SyncroSim results scenario.
}
\examples{
# Update an old scenario to allow rsyncrosim to access spatial output
multiband(myResultScenario,action="rebuild")

# Combine spatial outputs into multi-band rasters containing a layer for each timetep.
multiband(myResultScenario,action="apply",grouping="Timestep")

# Combine spatial outputs into multi-band rasters containing a layer for each iteration.
multiband(myResultScenario,action="apply",grouping="Iteration")

# Combine spatial outputs into multi-band rasters containing a layer for each timestep and iteration.
multiband(myResultScenario,action="apply",grouping="All")

# Remove multi-banding
multiband(myResultScenario,action="remove")

}

