% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{name<-}
\alias{name<-}
\title{Set the project or scenario names.}
\usage{
name(ssimObject) <- value
}
\arguments{
\item{ssimObject}{Scenario/Project}

\item{value}{The new name.}
}
\description{
Set the name of a SyncroSim Project or Scenario.
}

