% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{methods}
\name{package}
\alias{package}
\alias{package,missingOrNULL-method}
\alias{package,character-method}
\alias{package,Session-method}
\title{Installed packages}
\usage{
package(session)

\S4method{package}{missingOrNULL}(session)

\S4method{package}{character}(session)

\S4method{package}{Session}(session)
}
\arguments{
\item{session}{Session.}
}
\value{
A dataframe of packages
}
\description{
Packages installed with this version of SyncroSim
}
