% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/session.R, R/ssimLibrary.R
\docType{methods}
\name{path}
\alias{path}
\alias{path,SSimLibrary-method}
\alias{path,Session-method}
\title{The path to an object on disk}
\usage{
path(x)

\S4method{path}{Session}(x)

\S4method{path}{SSimLibrary}(x)
}
\arguments{
\item{x}{An object containing a path.}
}
\description{
The path to a SyncroSim Session, SSimLibarary, Project or Scenario on disk.
}
\section{Methods (by class)}{
\itemize{
\item \code{Session}: Path to the Syncrosim console in a Session.

\item \code{SSimLibrary}: Path to an SSimLibrary on disk.
}}

