% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{projects}
\alias{projects}
\title{The projects in a SyncroSim library.}
\usage{
projects(x)
}
\arguments{
\item{x=}{An SSimLibrary object}
}
\value{
A list of projects identified by the project id. Each element of the list contains a SyncroSim Project object.
}
\description{
Get a list of projects in a SyncroSim library.
}
\examples{
myProjects = projects(ssimLibrary(model="stsim",name="stsim"))
}

