% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimLibrary.R
\name{removeProject}
\alias{removeProject}
\title{Remove projects or projects from a Library}
\usage{
removeProject(ssimObject, project = NULL, force = F, forceElements = F)
}
\arguments{
\item{ssimObject}{SsimLibrary/Project or path to a library.}

\item{project}{character, numeric, or vector of these. One or more project names or ids. If NULL all projects in ssimObject will be removed.}

\item{force}{logical. If FALSE (default), user will be prompted to approve removal of each project.}

\item{forceElements}{logical. If TRUE then returns a single removal request as a named list; otherwise returns a single removal request as a character string.}
}
\value{
A "saved" or failure message, or a list of these for each project.
}
\description{
Removes one or more projects from a SyncroSim library. Note this is irreversable.
}
\examples{
TODO – update examples
myLibrary = ssimLibrary(session=devSession)
myProject = project(myLibrary,project="a project")
project(myLibrary)
removeProject(myLibrary,project="a project")
project(myLibrary)

}

