% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimLibrary.R
\name{removeScenario}
\alias{removeScenario}
\title{Remove scenario or scenarios from library}
\usage{
removeScenario(ssimObject, scenario = NULL, force = FALSE,
  forceElements = FALSE)
}
\arguments{
\item{ssimObject}{SsimLibrary/Project/Scenario or path to a library.}

\item{scenario}{NULL, character, numeric, or vector of these. One or more scenario names or ids. If NULL all scenarios in ssimObject will be removed.}

\item{force}{logical. If FALSE (default) user is prompted to approve each removal.}

\item{forceElements}{logical. If TRUE then returns a single removal request as a named list; otherwise returns a single removal request as a character string.}
}
\value{
A "saved" or failure message, or a list of these for each scenario.
}
\description{
Removes one or more scenarios from a SyncroSim library. Note that this is irreversable.
}
\examples{
TODO – update examples
myLibrary = ssimLibrary()
myScenario = scenario(project(myLibrary),scenario="Scenario")
scenario(myLibrary)
removeScenario(myLibrary,scenario="Scenario")
scenario(myLibrary)
}

