% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{reportProgress}
\alias{reportProgress}
\title{Reports progress for a SyncroSim simulation}
\usage{
reportProgress(iteration, timestep)
}
\arguments{
\item{iteration}{integer.  The current iteration.}

\item{timestep}{integer.  The current timestep.}
}
\description{
Reports progress for a SyncroSim simulation
}
