% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveFolderData.R
\name{retrieveFolderData}
\alias{retrieveFolderData}
\alias{retrieveFolderData,character-method}
\alias{retrieveFolderData,SsimObject-method}
\title{Retrieve folder data for a SsimLibrary}
\usage{
retrieveFolderData(ssimLibrary)

\S4method{retrieveFolderData}{character}(ssimLibrary)

\S4method{retrieveFolderData}{SsimObject}(ssimLibrary)
}
\arguments{
\item{ssimLibrary}{\code{\link{SsimLibrary}}}
}
\value{
Returns a \code{data.frame} with information about project folders in the 
SsimLibrary
}
\description{
Retrieve a table of folder data for a \code{\link{SsimLibrary}}. 
The folder data consists of folder names and corresponding folder IDs
that will be used to organize Scenarios in the SyncroSim User Interface.
}
\examples{
\donttest{
# Specify file path and name of new SsimLibrary
myLibraryName <- file.path(tempdir(), "testlib")

# Set up a SyncroSim Session and SsimLibrary
mySession <- session()
myLibrary <- ssimLibrary(name = myLibraryName, session = mySession)

# Retrieve folder data from the SsimLibrary
retrieveFolderData(myLibrary)
}

}
