% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsyncrosim.R
\docType{package}
\name{rsyncrosim}
\alias{rsyncrosim}
\title{rsyncrosim: The R interface to SyncroSim: https://syncrosim.com/}
\description{
rsyncrosim provides an interface to SyncroSim, a generalized framework for
running and managing scenario-based stochastic simulations over space and time.
Different kinds of simulation models can "plug-in" to SyncroSim as packages and
take advantage of general features common to many kinds of simulation models,
such as defining scenarios of inputs, running Monte Carlo simulations, and
viewing charts and maps of outputs.
}
\details{
To learn more about rsyncrosim, start with the vignette tutorial:
\code{browseVignettes("rsyncrosim")}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://syncrosim.github.io/rsyncrosim/}
  \item Report bugs at \url{https://github.com/syncrosim/rsyncrosim/issues/}
}

}
\author{
\strong{Maintainer}: ApexRMS \email{rsyncrosim@syncrosim.com} [copyright holder]

Authors:
\itemize{
  \item Colin Daniel \email{colin.daniel@apexrms.com}
  \item Josie Hughes \email{josie.hughes@canada.ca}
  \item Valentin Lucet
  \item Alex Embrey
  \item Katie Birchard \email{katie.birchard@apexrms.com}
  \item Leonardo Frid
  \item Tabitha Kennedy
  \item Shreeram Senthivasan
  \item Vicki Zhang \email{vicki.zhang@apexrms.com}
  \item Hannah Adams \email{hannah.adams@apexrms.com}
}

}
