% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{runtimeDataFolder}
\alias{runtimeDataFolder}
\title{SyncroSim Data Folder}
\usage{
runtimeDataFolder(scenario, datasheetName)
}
\arguments{
\item{scenario}{\code{\link{Scenario-class}} object. A SyncroSim result Scenario}

\item{datasheetName}{character. The datasheet name}
}
\value{
Returns a data folder name for the specified datasheet.
}
\description{
This function is part of a set of functions designed to facilitate the
development of R-based Syncrosim Packages. This function creates and returns
a SyncroSim Data Folder.
}
\examples{
\dontrun{
dataFolder <- runtimeDataFolder()
}

}
