% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{runtimeOutputFolder}
\alias{runtimeOutputFolder}
\title{SyncroSim DataSheet Output Folder}
\usage{
runtimeOutputFolder(scenario, datasheetName)
}
\arguments{
\item{scenario}{\code{\link{Scenario}} object. A SyncroSim result Scenario}

\item{datasheetName}{character. The output Datasheet name}
}
\value{
Returns a folder name for the specified datasheet.
}
\description{
This function is part of a set of functions designed to facilitate the
development of R-based Syncrosim Packages. This function creates and returns 
a SyncroSim DataSheet Output Folder.
}
\examples{
\dontrun{
outputFolder <- runtimeOutputFolder()
}

}
