% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{session<-}
\alias{session<-}
\title{Set a SyncroSim session.}
\usage{
session(x) <- value
}
\arguments{
\item{x}{An SsimObject.}

\item{value}{A SyncroSim Session.}
}
\value{
An SyncroSim object containing a Session.
}
\description{
Set the Session of a SsimLibrary, Project or Scenario object.
}
\examples{
myLibrary = ssimLibrary()
session(myLibrary)=session()
session(myLibrary)
}

