% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointSession.R
\name{setBreakpoint}
\alias{setBreakpoint}
\title{Set breakpoint of a Scenario.}
\usage{
setBreakpoint(x, breakpointType, transformerName, arguments, callback)
}
\arguments{
\item{x}{A SyncroSim Scenario}

\item{breakpointType}{bi: before iteration; ai: after iteration; bt:before timestep; at: aftertimestep}

\item{transformerName}{'stsim:core-transformer' or?}

\item{arguments}{A vector of timesteps or iterations e.g. c(1,2)}

\item{callback}{The function to apply. See STSimBreakpointsTutorial.R for details.}
}
\value{
An SyncroSim Scenario object containing breakpoints
}
\description{
Add a Breakpoint object to breakpoints of a Scenario.
}

