% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{setProperties}
\alias{setProperties}
\title{Set the properties of a scenario.}
\usage{
setProperties(x, author = NULL, description = NULL, readOnly = NULL)
}
\arguments{
\item{x}{An Scenario object.}

\item{author}{An author name.}

\item{description}{A description.}

\item{readOnly}{TRUE or FALSE.}
}
\value{
"saved" or a failure message
}
\description{
Set the author, description and/or readOnly status of a scenario.
}

