% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setRows.R
\name{setRows}
\alias{setRows}
\title{Set values in a dataframe.}
\usage{
setRows(x, conditions, values)
}
\arguments{
\item{x}{A dataframe.}

\item{conditions}{A list of conditions of the form list(colName1=colValue1,colName2=colValue2)}

\item{values}{A list of values of the form list(colName1=colValue1,colName2=colValue2)}
}
\value{
A dataframe containing the values.
}
\description{
Modifies values in one or more rows of a data frame.
Values are assigned to all rows that meet the conditions.
}

