% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signin.R
\name{signIn}
\alias{signIn}
\title{Signs in to SyncroSim}
\usage{
signIn(session = NULL)
}
\arguments{
\item{session}{\code{\linkS4class{Session}} object. If \code{NULL}(default), the
default session will be used}
}
\value{
Character string: whether sign in was successful or not.
}
\description{
Signs in to syncrosim.com to authenticate user credentials.
}
\examples{
\dontrun{
# Sign in to SyncroSim session
mySession <- session()
signIn(mySession)
}

}
