% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signout.R
\name{signOut}
\alias{signOut}
\title{Signs out of SyncroSim}
\usage{
signOut(session = NULL)
}
\arguments{
\item{session}{\code{\link{Session-class}} object. If \code{NULL}(default), the
default session will be used}
}
\value{
Character string: whether sign out was successful or not.
}
\description{
Signs out of syncrosim.com.
}
\examples{
\dontrun{
# Sign out of SyncroSim session
mySession <- session()
signOut(mySession)
}

}
