% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silent.R
\docType{methods}
\name{silent<-}
\alias{silent<-}
\alias{silent<-,Session-method}
\alias{silent<-,character-method}
\title{Set silent property of a Session}
\usage{
silent(session) <- value

\S4method{silent}{character}(session) <- value

\S4method{silent}{Session}(session) <- value
}
\arguments{
\item{session}{Session}

\item{value}{logical}
}
\description{
Set silent property of a sessio to TRUE or FALSE
}

