% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialData.R
\name{spatialData}
\alias{spatialData}
\title{Get spatial inputs or outputs from a SyncroSim scenario.}
\usage{
spatialData(x, sheet, iterations = NULL, timesteps = NULL,
  nameFilters = NULL, rat = NULL)
}
\arguments{
\item{x}{A SyncroSim results Scenario or list of SyncroSim result Scenarios.}

\item{sheet}{The name of a spatial datasheet.}

\item{iterations}{A vector of iterations. If NULL(default) all available iterations will be included}

\item{timesteps}{A vector of timesteps. If NULL(default) all available timesteps will be included.}

\item{nameFilters}{A vector of strings. Only layer name that include these terms will be returned.}

\item{rat}{An (optional) raster attribute table. This is dataframe with ID, (optional) Color, and other columns. See raster::ratify() for details.}
}
\value{
A RasterStack or RasterBrick object. See raster package documentation for details.
}
\description{
Get spatial inputs or outputs from a SyncroSim scenario.
}
\details{
The Color column of a rat table should have one of these formats:
\itemize{
  \item {R,G,B,alpha: } {4 numbers representing red, green, blue and alpha, separated by commas, and scaled between 0 and 255. See rgb() for details.}
  \item {R colour names: } {See colors() for options.}
  \item {hexadecimal colors: } {As returned by R functions such as rainbow(), heat.colors(), terrain.colors(), topo.colors(), gray(), etc.}
}

The names() of the returned raster stack contain metadata.
For datasheets without Filename this is: paste0(<datasheet name>,".Scn",<scenario id>,".",<tif name>)
For datasheets containing Filename this is: paste0(<datasheet name>,".Scn",<scenario id>,".It",<iteration>,".Ts",<timestep>)
}

