% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlStatements.R
\name{sqlStatements}
\alias{sqlStatements}
\title{Get SELECT and GROUP BY Statements}
\usage{
sqlStatements(groupBy = NULL, aggregate = NULL, aggregateFunction = "SUM")
}
\arguments{
\item{groupBy}{Vector of variables to GROUP BY.}

\item{aggregate}{Vector of variables to aggregate using aggregateFunction}

\item{aggregateFunction}{An SQL aggregate function (e.g. SUM, COUNT)}
}
\value{
A list of SQL SELECT and GROUP BY statements.
}
\description{
Creates SELECT and GROUP BY SQL Satements.
Variables are column names.
Variables not included in groupBy or aggregate will be dropped from the table.
}

