% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{ssimEnvironment}
\alias{ssimEnvironment}
\title{SyncroSim Environment.}
\usage{
ssimEnvironment()
}
\arguments{
\item{NULL}{}
}
\value{
A data.frame of SyncroSim specific environment variables.
}
\description{
Retrieves SyncroSim specific environment variables.
}
