% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{ssimEnvTempFolder}
\alias{ssimEnvTempFolder}
\title{SyncroSim Temporary Folder}
\usage{
ssimEnvTempFolder(folderName)
}
\arguments{
\item{folderName}{character.  The folder name}
}
\value{
a temporary folder name
}
\description{
Creates and returns a SyncroSim Temporary Folder
}
