% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimRatify.R
\docType{methods}
\name{ssimRatify<-}
\alias{ssimRatify<-}
\alias{ssimRatify<-,RasterLayer-method}
\title{Set attributes and colors of a RasterLayer object.}
\usage{
ssimRatify(raster) <- value

\S4method{ssimRatify}{RasterLayer}(raster) <- value
}
\arguments{
\item{raster}{A RasterLayer.}

\item{rat}{dataframe. A raster attribute table is a dataframe with ID, (optional) Color, and other columns. See raster::ratify() for details.}
}
\description{
Set attributes and colors of a Raster object.
This is a wrapper around ratify() function from the raster package. 
The optional Color column accepts names, hexadecimal colors, and RGB color strings exported by SyncroSim.
These are converted to hexadecimal colors in the hexColor column of the raster attribute table.
}
\details{
The (optional) Color column of a rat table should have one of these formats:
\itemize{
  \item {alpha,R,G,B: } {4 numbers representing red, green, blue and alpha, separated by commas, and scaled between 0 and 255. See rgb() for details.}
  \item {R colour names: } {See colors() for options.}
  \item {hexadecimal colors: } {As returned by R functions such as rainbow(), heat.colors(), terrain.colors(), topo.colors(), gray(), etc.}
}
}
\examples{

levels(myRaster) #retrieve raster attribute table

}

