% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{ssimTempFolder}
\alias{ssimTempFolder}
\title{SyncroSim Temporary Folder}
\usage{
ssimTempFolder(scenario, subFolderName)
}
\arguments{
\item{scenario}{Scenario.  A SyncroSim result scenario.}

\item{subFolderName}{character.  A subfolder name.}
}
\value{
a temporary folder name
}
\description{
Creates and returns a SyncroSim Temporary Folder
}
