% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimLibrary.R
\name{ssimUpdate}
\alias{ssimUpdate}
\title{Apply updates.}
\usage{
ssimUpdate(x)
}
\arguments{
\item{x}{An SsimLibrary object, or a Project or Scenario associated with a Library}
}
\value{
"saved" or a failure message from the console.
}
\description{
Apply updates to a SyncroSim Library.
}

