% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{updateRunLog}
\alias{updateRunLog}
\title{Function to write to the SyncroSim run log}
\usage{
updateRunLog(..., sep = "")
}
\arguments{
\item{...}{One or more objects which can be coerced to character
which are pasted together using `sep`.}

\item{sep}{character.  Used to separate terms. Not NA_character_}
}
\value{
No returned value, used for side effects
}
\description{
This function is designed to facilitate the development of R-based Syncrosim 
Packages by allowing developers to send messages to the run log.
}
\examples{
\dontrun{
# Write a message to run log
updateRunLog(msg)

# Construct and write a message to run log
updateRunLog(msg, additionalMsg, sep = " ")
}

}
