% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addModule.R
\name{addModule}
\alias{addModule}
\alias{addModule,character-method}
\title{Add module}
\usage{
addModule(filename, session = NULL)

\S4method{addModule}{character}(filename, session = NULL)
}
\arguments{
\item{filename}{Character string or vector of these. The path to an .ssimpkg file on disk, or a vector of filepaths.}

\item{session}{Session.}
}
\description{
Add module or modules to SyncroSim
Deprecated.  See: \code{\link{addPackage}} and \code{\link{addPackageFile}}
}
