% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{addModule}
\alias{addModule}
\title{Add modules}
\usage{
addModule(filename, session = NULL)
}
\arguments{
\item{filename}{The path to an .ssimpkg file on disk, or a vector of filepaths}

\item{session}{A SyncroSim \code{\link{Session}} object.}
}
\description{
Add module or modules to SyncroSim
}

