% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{addModule}
\alias{addModule}
\title{Add module}
\usage{
addModule(filename, session = NULL)
}
\arguments{
\item{filename}{character string or vector of these. The path to an .ssimpkg
file on disk, or a vector of filepaths.}

\item{session}{\code{\link{Session}} object}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Please use \code{\link{installPackage}} instead.
}
\keyword{internal}
