% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{addModules<-}
\alias{addModules<-}
\title{Add modules}
\usage{
addModules(x) <- value
}
\arguments{
\item{value}{The path to an .ssimpkg file on disk, or a vector of filepaths}

\item{object}{A SyncroSim \code{\link{Session}} object.}
}
\description{
Add module or modules to this version of SyncroSim
TO DO: Doesn't currently do anything. Need console commands for adding and removing modules.
}

