% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addon.R
\name{addon}
\alias{addon}
\alias{addon,character-method}
\alias{addon,missingOrNULL-method}
\alias{addon,Session-method}
\alias{addon,SsimObject-method}
\title{addon(s) of an SsimLibrary or Session}
\usage{
addon(ssimObject)

\S4method{addon}{character}(ssimObject)

\S4method{addon}{missingOrNULL}(ssimObject)

\S4method{addon}{Session}(ssimObject)

\S4method{addon}{SsimObject}(ssimObject)
}
\arguments{
\item{ssimObject}{SsimLibrary/Project/Scenario or Session.}
}
\value{
A dataframe of addons.
}
\description{
The addon(s) of an SsimLibrary or Session.
}
\examples{
\donttest{
addon(ssimLibrary(name = "mylib"))
}

}
