% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimLibrary.R
\name{addons}
\alias{addons}
\title{addons of an SSimLibrary}
\usage{
addons(x, all = F)
}
\arguments{
\item{x}{An SSimLibrary, or a Project/Scenario object associated with an SSimLibrary.}

\item{all}{If T, all available addons are returned. Otherwise, only enabled addons.}
}
\value{
A dataframe of addons.
}
\description{
The addons of an SSimLibrary.
}
\examples{
addons(ssimLibrary(model="stsim",name="stsim"))
}

