% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addons.R
\name{addons}
\alias{addons}
\title{addons of an SsimLibrary or Session}
\usage{
addons(ssimObject, all = F)
}
\arguments{
\item{ssimObject}{SsimLibrary/Project/Scenario or Session.}

\item{all}{logical. If TRUE, all available addons are returned. Otherwise, only enabled addons.}
}
\value{
A dataframe of addons.
}
\description{
The addons of an SsimLibrary or Session.
}
\examples{
addons(ssimLibrary(name="stsim"))
}

