% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimLibrary.R
\name{addons}
\alias{addons}
\title{addons of an SsimLibrary or Session}
\usage{
addons(ssimObject, all = F)
}
\arguments{
\item{ssimObject}{An SsimLibrary, or a Project/Scenario object associated with an SsimLibrary.}

\item{all}{If T, all available addons are returned. Otherwise, only enabled addons.}
}
\value{
A dataframe of addons.
}
\description{
The addons of an SsimLibrary or Session.
}
\examples{
addons(ssimLibrary(name="stsim"))
}

