% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPackage.R
\docType{methods}
\name{addPackage}
\alias{addPackage}
\alias{addPackage,character-method}
\title{Adds a package to SyncroSim}
\usage{
addPackage(name = NULL, session = NULL)

\S4method{addPackage}{character}(name = NULL, session = NULL)
}
\arguments{
\item{name}{Character string.  The name of the package to install from the online package server.}

\item{session}{Session.}
}
\description{
Adds a package to SyncroSim
}
