% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPackage.R
\name{addPackage}
\alias{addPackage}
\alias{addPackage,ANY,character-method}
\alias{addPackage,ANY,missingOrNULL-method}
\alias{addPackage,ANY,Session-method}
\title{Adds a package to SyncroSim}
\usage{
addPackage(name, session = NULL)

\S4method{addPackage}{ANY,character}(name, session = NULL)

\S4method{addPackage}{ANY,missingOrNULL}(name, session = NULL)

\S4method{addPackage}{ANY,Session}(name, session = NULL)
}
\arguments{
\item{name}{Character string.  The name of the package to install.}

\item{session}{Session.}
}
\value{
This function will invisibly return `TRUE` upon success (i.e.successful 
install) and `FALSE` upon failure.
}
\description{
Adds a package to SyncroSim. This functions will query the SyncroSim 
package server for the package name provided as input.
}
