% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPackageFile.R
\name{addPackageFile}
\alias{addPackageFile}
\alias{addPackageFile,ANY,character-method}
\alias{addPackageFile,ANY,missingOrNULL-method}
\alias{addPackageFile,ANY,Session-method}
\title{Adds a package to SyncroSim}
\usage{
addPackageFile(filename, session = NULL)

\S4method{addPackageFile}{ANY,character}(filename, session = NULL)

\S4method{addPackageFile}{ANY,missingOrNULL}(filename, session = NULL)

\S4method{addPackageFile}{ANY,Session}(filename, session = NULL)
}
\arguments{
\item{filename}{Character string.  The path to a SyncroSim package file.}

\item{session}{A \code{\link{Session}} object.}
}
\value{
This function invisibly returns `TRUE` upon success (i.e.successful 
install) and `FALSE` upon failure.
}
\description{
This function adds a package to SyncroSim from a package file (ends in .ssimpkg).
The list of SyncroSim packages can be found \href{https://syncrosim.com/packages/}{here}.
}
\examples{
\donttest{
temp_dir <- tempdir()
myses <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"), session = myses)

addPackageFile(myLibrary, "path/to/package_file.ssimpkg")
}

}
\seealso{
\link{addPackage}
}
