% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{addPackageFile}
\alias{addPackageFile}
\title{Adds a package to SyncroSim}
\usage{
addPackageFile(filename, session = NULL)
}
\arguments{
\item{filename}{character string.  The path to a SyncroSim package file}

\item{session}{\code{\link{Session}} object}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Please use \code{\link{addPackage}} instead.
}
\keyword{internal}
