% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPackageFile.R
\name{addPackageFile}
\alias{addPackageFile}
\alias{addPackageFile,ANY,character-method}
\alias{addPackageFile,ANY,missingOrNULL-method}
\alias{addPackageFile,ANY,Session-method}
\title{Adds a package to SyncroSim}
\usage{
addPackageFile(filename, session = NULL)

\S4method{addPackageFile}{ANY,character}(filename, session = NULL)

\S4method{addPackageFile}{ANY,missingOrNULL}(filename, session = NULL)

\S4method{addPackageFile}{ANY,Session}(filename, session = NULL)
}
\arguments{
\item{filename}{Character string.  The path to a SyncroSim package file.}

\item{session}{\code{\link{Session}}.}
}
\value{
This function invisibly returns `TRUE` upon success (i.e.successful 
install) and `FALSE` upon failure.
}
\description{
Adds a package to SyncroSim from a package file.
}
