% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogentags.R
\name{autogentags<-}
\alias{autogentags<-}
\alias{autogentags<-,character-method}
\alias{autogentags<-,Scenario-method}
\title{Auto Generation Tags for a Scenario.}
\usage{
autogentags(ssimObject, tags) <- value

\S4method{autogentags}{character}(ssimObject, tags) <- value

\S4method{autogentags}{Scenario}(ssimObject, tags) <- value
}
\arguments{
\item{ssimObject}{Scenario}

\item{tags}{character}
}
\value{
The updated ssimObject.
}
\description{
Sets the Auto Generation Tags for a Scenario.
}
