% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\docType{methods}
\name{backup}
\alias{backup}
\alias{backup,SsimObject-method}
\alias{backup,character-method}
\title{Backup an SsimLibrary.}
\usage{
backup(ssimObject)

\S4method{backup}{character}(ssimObject)

\S4method{backup}{SsimObject}(ssimObject)
}
\arguments{
\item{ssimObject}{SsimLibrary/Project/Scenario.}
}
\description{
Backup an SsimLibrary.
}

