% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\name{backup}
\alias{backup}
\alias{backup,character-method}
\alias{backup,SsimObject-method}
\title{Backup a SsimLibrary.}
\usage{
backup(ssimObject)

\S4method{backup}{character}(ssimObject)

\S4method{backup}{SsimObject}(ssimObject)
}
\arguments{
\item{ssimObject}{SsimLibrary/Project/Scenario.}
}
\value{
This function invisibly returns `TRUE` upon success (i.e.successful 
backup) and `FALSE` upon failure.
}
\description{
Backup a SsimLibrary.
}
