% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{basePackage}
\alias{basePackage}
\title{Installed base packages}
\usage{
basePackage(ssimObject = NULL)
}
\arguments{
\item{ssimObject}{An object of class \code{\link{Session}} or \code{\link{SsimLibrary}}.}
}
\value{
A dataframe of base packages (for Session) or named vector of character strings (for SsimLibrary).
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
Please use \code{\link{package}} instead.
}
