% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basePackage.R
\name{basePackage}
\alias{basePackage}
\alias{basePackage,character-method}
\alias{basePackage,missingOrNULL-method}
\alias{basePackage,Session-method}
\alias{basePackage,SsimLibrary-method}
\title{Installed base packages}
\usage{
basePackage(ssimObject = NULL)

\S4method{basePackage}{character}(ssimObject = NULL)

\S4method{basePackage}{missingOrNULL}(ssimObject = NULL)

\S4method{basePackage}{Session}(ssimObject = NULL)

\S4method{basePackage}{SsimLibrary}(ssimObject = NULL)
}
\arguments{
\item{ssimObject}{An object of type \code{\link{Session}} or \code{\link{SsimLibrary}}.}
}
\value{
A dataframe of base packages (for Session) or named vector of character strings (for SsimLibrary).
}
\description{
This retrieves the Base packages installed with this version of SyncroSim.
The list of SyncroSim packages can be found \href{https://syncrosim.com/packages/}{here}.
}
\examples{
\donttest{
temp_dir <- tempdir()
mySession <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"), session = mySession)

basePackage(myLibrary)
}

}
