% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basePackage.R
\name{basePackage}
\alias{basePackage}
\alias{basePackage,character-method}
\alias{basePackage,missingOrNULL-method}
\alias{basePackage,Session-method}
\alias{basePackage,SsimLibrary-method}
\title{Installed base packages}
\usage{
basePackage(ssimObject = NULL)

\S4method{basePackage}{character}(ssimObject = NULL)

\S4method{basePackage}{missingOrNULL}(ssimObject = NULL)

\S4method{basePackage}{Session}(ssimObject = NULL)

\S4method{basePackage}{SsimLibrary}(ssimObject = NULL)
}
\arguments{
\item{ssimObject}{Session or SsimLibrary.}
}
\value{
A dataframe of base packages (for Session) or named vector of character strings (for SsimLibrary).
}
\description{
Base packages installed with this version of SyncroSim.
}
