% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint.R
\docType{methods}
\name{breakpoint}
\alias{breakpoint}
\alias{breakpoint,Scenario-method}
\title{Lists the breakpoints for a Scenario.}
\usage{
breakpoint(x)

\S4method{breakpoint}{Scenario}(x)
}
\arguments{
\item{x}{A SyncroSim Scenario}
}
\description{
Lists the breakpoints for a Scenario.
}
