% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAClassDefinitions.R
\docType{class}
\name{Chart-class}
\alias{Chart-class}
\title{SyncroSim Chart class}
\description{
\code{Chart} object representing a SyncroSim Chart object. A Chart object
is used to create line or column charts from tabular output data in the
and can be viewed using the SyncroSim User Interface.
}
\section{Slots}{

\describe{
\item{\code{session}}{\code{\link{Session-class}} object. The Session associated with the
Chart's SsimLibrary}

\item{\code{filepath}}{character string. The path to the Chart's SsimLibrary on disk}

\item{\code{chartId}}{integer. The Chart id}

\item{\code{projectId}}{integer. The Project id}
}}

\seealso{
See \code{\link{chart}} for options when creating or loading a
SyncroSim Chart
}
