% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chart}
\alias{chart}
\title{Create or open a chart}
\usage{
chart(
  ssimObject = NULL,
  chart = NULL,
  type = "line",
  create = FALSE,
  summary = FALSE
)
}
\arguments{
\item{ssimObject}{\code{\link{Project}} or \code{\link{Scenario}} object}

\item{chart}{character or integer. If character, then will either open an
existing chart if \code{create=FALSE}, or will create a new chart with the
given name if the chart does not exist yet or \code{create=TRUE}.
If integer, will open the existing chart with the given chart ID (if the
ID exists). If no value is provided, the default is to create a new chart
called "_Chart"}

\item{type}{character. Can be "line" (Default) or "bar"}

\item{create}{logical. Whether to create a new chart if the chart name given
already exists in the SyncroSim library. If \code{FALSE} (Default), then will
return the existing chart with the given name. If \code{TRUE}, then will
return a new chart with the same name as an existing chart (but different
chart ID)}

\item{summary}{logical. If \code{TRUE}, returns a summary of chart
information as an R data.frame. If \code{FALSE} (Default), then returns
a SyncroSim Chart object}
}
\value{
A \code{Chart} object representing a SyncroSim chart
}
\description{
Create or open a \code{\link{Chart}} from a SyncroSim
\code{\link{Project}}.
}
