% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartData.R
\name{chartData}
\alias{chartData}
\alias{chartData,Chart-method}
\title{Sets the \code{\link{Chart}} type to "Line"}
\usage{
chartData(
  chart,
  type = "Line",
  x = NULL,
  y = NULL,
  timesteps = NULL,
  iterationType = "Mean",
  iteration = 1
)

\S4method{chartData}{Chart}(
  chart,
  type = "Line",
  x = NULL,
  y = NULL,
  timesteps = NULL,
  iterationType = "Mean",
  iteration = 1
)
}
\arguments{
\item{chart}{\code{\link{Chart}} object}

\item{type}{character. Chart type. Can be "Line" (Default) or "Column".}

\item{x}{character or character vector. X variable(s) to plot. If \code{NULL}
(Default), then will plot timesteps along the X axis.}

\item{y}{character or character vector. Y variable(s) to plot.}

\item{timesteps}{integer vector. The range of timesteps to plot against
(Default is the minimum and maximum timesteps defined in the simulation).}

\item{iterationType}{character. How to display multiple iterations in the
chart. Can be "Mean" (Default), "Single", or "All".}

\item{iteration}{integer. If the \code{iterationType} is set to "Single", this argument
determines which iteration to display. Default is 1.}
}
\value{
A \code{Chart} object representing a SyncroSim chart
}
\description{
Sets the \code{\link{Chart}} type to "Line" and adds the variables to plot
in the line chart.
}
\examples{
\dontrun{
# Create a chart object
myChart <- chart(myProject, chart = "New Chart")

# Set the chart type and data
myChart <- chartData(myChart, y = c("variable1", "variable2"),
timesteps = c(0,10), iterationType = "single", iteration = 1)

}

}
