% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{command}
\alias{command}
\title{SyncroSim console command}
\usage{
command(args, cSession = NULL, silent = F)
}
\arguments{
\item{args}{A list of arguments to the SyncroSim console.}

\item{cSession}{A session object. If NULL, a default session will be used.}

\item{silent}{If TRUE, warnings from the console are ignored. Otherwise they are printed.}
}
\value{
Output from the SyncroSim console.
}
\description{
\code{command} issues a command to the SyncroSim console and returns the output.
}
\examples{
#Use a default session to creat a new library
args = list(create=NULL,library=NULL,name=paste0(getwd(),"/temp.ssim",model="stsim:model-transformer")
output = command(args)
output
}

