% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{command}
\alias{command}
\title{SyncroSim console command}
\usage{
command(args, session = NULL, printCmd = F,
  program = "/SyncroSim.Console.exe", silent = NULL, wait = T)
}
\arguments{
\item{args}{A list of arguments to the SyncroSim console.}

\item{session}{A SyncroSim session object. If NULL, a default session will be used.}

\item{printCmd}{If T, the command string is printed.}

\item{silent}{If NULL (default) use session@silent. If T suppress warnings from console.}

\item{wait}{If TRUE (default) R will wait for the command to finish before proceeding.}
}
\value{
Output from the SyncroSim console.
}
\description{
\code{command} issues a command to the SyncroSim console and returns the output.
}
\examples{
# Use a default session to creat a new library
args = list(create=NULL,ssimLibrary=NULL,name=paste0(getwd(),"/temp.ssim",model="stsim:model-transformer")
output = command(args)
output
}

