% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{datasheet}
\alias{datasheet}
\title{datasheet}
\usage{
datasheet(x, ...)
}
\arguments{
\item{x}{An SSimLibrary, Project or Scenario object. Or the path to a library on disk.}

\item{name}{The sheet name}

\item{project}{Project name or id.}

\item{scenario}{Scenario name or id.}

\item{optional}{If FALSE (default) returns only required columns. If TRUE returns optional columns also.}

\item{empty}{If FALSE (default) returns data (if any). If FALSE returns empty dataframe of correct format.}
}
\value{
A dataframe.
}
\description{
Gets Syncrosim datasheet.
}
\details{
If !empty & !optional: return required columns and columns containing data
}
\examples{

}

