% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{datasheet}
\alias{datasheet}
\title{Get a datasheet}
\usage{
datasheet(x, name, project = NULL, scenario = NULL, optional = F,
  empty = F, sheetNames = NULL, checkDependencies = T)
}
\arguments{
\item{x}{An SSimLibrary, Project or Scenario object. Or the path to a library on disk.}

\item{name}{The sheet name}

\item{project}{Project name or id.}

\item{scenario}{Scenario name or id.}

\item{optional}{If FALSE (default) returns only required columns. If TRUE returns optional columns also.}

\item{empty}{If FALSE (default) returns data (if any). If FALSE returns empty dataframe of correct format.}

\item{sheetNames}{Output from datasheets(). Used internally to speed calculation of dependencies.}

\item{checkDependencies}{If TRUE (default) dependencies are checked. Set to FALSE to speed calculations.}
}
\value{
A dataframe.
}
\description{
Gets Syncrosim datasheet.
}
\details{
If !empty & !optional: return required columns and columns containing data
}
\examples{

}

