% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasheetRaster.R
\docType{methods}
\name{datasheetRaster}
\alias{datasheetRaster}
\alias{datasheetRaster,Scenario-method}
\alias{datasheetRaster,SsimObject-method}
\alias{datasheetRaster,list-method}
\title{Get spatial inputs or outputs from a Scenario(s).}
\usage{
datasheetRaster(ssimObject, datasheet, column = NULL, scenario = NULL,
  iteration = NULL, timestep = NULL, subset = NULL, rat = NULL,
  forceElements = F)

\S4method{datasheetRaster}{list}(ssimObject, datasheet, column = NULL,
  scenario = NULL, iteration = NULL, timestep = NULL, subset = NULL,
  rat = NULL, forceElements = F)

\S4method{datasheetRaster}{SsimObject}(ssimObject, datasheet, column = NULL,
  scenario = NULL, iteration = NULL, timestep = NULL, subset = NULL,
  rat = NULL, forceElements = F)

\S4method{datasheetRaster}{Scenario}(ssimObject, datasheet, column = NULL,
  scenario = NULL, iteration = NULL, timestep = NULL, subset = NULL,
  rat = NULL, forceElements = F)
}
\arguments{
\item{ssimObject}{SsimLibrary/Project/Scenario or list of Scenarios. If SsimLibrary/Project, then scenario argument is required.}

\item{datasheet}{character string. The name of the datasheet containing the raster data.}

\item{column}{character string. The name of the column in the datasheet containing the filenames for raster data. If NULL then use the first column that contains raster filenames.}

\item{scenario}{character string, integer, or vector of these. The scenarios to include. Required if ssimObject is an SsimLibrary/Project, ignored if ssimObject is a list of Scenarios.}

\item{iteration}{integer, character string, or vector of integer/character strings. Iteration(s) to include. If NULL then all iterations are included. If no Iteration column in the datasheet, then ignored.}

\item{timestep}{integer, character string, or vector of integer/character string. Timestep(s) to include. If NULL then all timesteps are included.  If no Timestep column in the datasheet, then ignored.}

\item{subset}{logical expression. logical expression indicating datasheet rows to return. e.g. expression(grepl("Ts0001",Filename,fixed=T)). See subset() for details.}

\item{rat}{An (optional) raster attribute table. This is dataframe with ID, (optional) Color, and other columns. See raster::ratify() for details.}

\item{forceElements}{logical. If TRUE then returns a single raster as a RasterStack; otherwise returns a single raster as a RasterLayer directly.}
}
\value{
A RasterLayer, RasterStack or RasterBrick object. See raster package documentation for details.
}
\description{
Get spatial inputs or outputs from one or more SyncroSim scenarios.
}
\details{
The Color column of a rat table should have one of these formats:
\itemize{
  \item {R,G,B,alpha: } {4 numbers representing red, green, blue and alpha, separated by commas, and scaled between 0 and 255. See rgb() for details.}
  \item {R colour names: } {See colors() for options.}
  \item {hexadecimal colors: } {As returned by R functions such as rainbow(), heat.colors(), terrain.colors(), topo.colors(), gray(), etc.}
}

The names() of the returned raster stack contain metadata.
For datasheets without Filename this is: paste0(<datasheet name>,".Scn",<scenario id>,".",<tif name>)
For datasheets containing Filename this is: paste0(<datasheet name>,".Scn",<scenario id>,".It",<iteration>,".Ts",<timestep>)
}
\examples{

datasheetRaster(myResult,datasheet="STSim_OutputSpatialState",
  subset=expression(grepl("Ts0001",Filename,fixed=T)))
}

