% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateModified.R
\name{dateModified}
\alias{dateModified}
\alias{dateModified,character-method}
\alias{dateModified,SsimLibrary-method}
\alias{dateModified,Project-method}
\alias{dateModified,Scenario-method}
\title{The last date a SsimLibrary, Project or Scenario was modified.}
\usage{
dateModified(ssimObject)

\S4method{dateModified}{character}(ssimObject)

\S4method{dateModified}{SsimLibrary}(ssimObject)

\S4method{dateModified}{Project}(ssimObject)

\S4method{dateModified}{Scenario}(ssimObject)
}
\arguments{
\item{ssimObject}{SsimLibrary, Project or Scenario.}
}
\value{
A character string of the date and time of the most recent modification 
to the ssimObject provided as input.
}
\description{
The most recent modification date of a \code{\link{SsimLibrary}}, \code{\link{Project}} or \code{\link{Scenario}}.
}
