% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateModified.R
\docType{methods}
\name{dateModified}
\alias{dateModified}
\alias{dateModified,Project-method}
\alias{dateModified,Scenario-method}
\alias{dateModified,SsimLibrary-method}
\title{The last date a SsimLibrary/Project/Scenario was modified.}
\usage{
dateModified(ssimObject)

\S4method{dateModified}{SsimLibrary}(ssimObject)

\S4method{dateModified}{Project}(ssimObject)

\S4method{dateModified}{Scenario}(ssimObject)
}
\arguments{
\item{ssimObject}{SsimLibrary/Project/Scenario.}
}
\description{
The most recent modification date of an SsimLibrary/Project/Scenario
}
\section{Methods (by class)}{
\itemize{
\item \code{SsimLibrary}: dateModified of an SsimLibrary.

\item \code{Project}: dateModified of a Project.

\item \code{Scenario}: dateModified of a Scenario.
}}

