% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateModified.R
\name{dateModified}
\alias{dateModified}
\title{The last date a SsimLibrary/Project/Scenario was modified.}
\usage{
dateModified(ssimObject)
}
\arguments{
\item{ssimObject}{SsimLibrary/Project/Scenario.}
}
\description{
The most recent modification date of an SsimLibrary/Project/Scenario
}

