% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{defaultModel<-}
\alias{defaultModel<-}
\title{Set defaultModel of a Session}
\usage{
defaultModel(session) <- value
}
\arguments{
\item{session}{Session}

\item{value}{character}
}
\description{
Set defaultModel of a session
}

