% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{defaultModel}
\alias{defaultModel}
\title{Get the default model from a \code{\link{Session}}.}
\usage{
defaultModel(session = NULL)
}
\arguments{
\item{session}{Session or character. A SyncroSim \code{\link{Session}} object or path to a session. If NULL, the default session will be used.}
}
\description{
Get the default model from a \code{\link{Session}}.
}

