% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultModel.R
\docType{methods}
\name{defaultModel}
\alias{defaultModel}
\alias{defaultModel,NULL-method}
\alias{defaultModel,Session-method}
\title{Get the default model from a Session.}
\usage{
defaultModel(session = NULL)

\S4method{defaultModel}{Session}(session = NULL)

\S4method{defaultModel}{`NULL`}(session = NULL)
}
\arguments{
\item{session}{Session or character. A Session object or path to a session. If NULL, the default session will be used.}
}
\value{
The default model of a Session.
}
\description{
Get the default model from a Session.
}

