% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete}
\alias{delete}
\title{Delete library, project, scenario, datasheet}
\usage{
delete(ssimObject, project = NULL, scenario = NULL, datasheet = NULL,
  force = F)
}
\arguments{
\item{ssimObject}{SsimLibrary/Project/Scenario, or path to a library.}

\item{project}{character, numeric, or vector of these. One or more project names or ids. Note that project argument is ignored if ssimObject is a list. Note that integer ids are slightly faster.}

\item{scenario}{character, numeric, or vector of these. One or more project names or ids. Note that scenario argument is ignored if ssimObject is a list. Note that integer ids are slightly faster.}

\item{datasheet}{character, numeric, or vector of these. One or more project names or ids.}

\item{force}{logical. If FALSE (default), user will be prompted to approve removal of each item.}
}
\value{
A list of "saved" or failure messages for each item.
}
\description{
Deletes one or more items. Note this is irreversable.
}
\examples{
TODO – update examples
myLibrary = ssimLibrary(session=devSession)
myProject = project(myLibrary,project="a project")
project(myLibrary)
removeProject(myLibrary,project="a project")
project(myLibrary)

}

