% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteLibrary.R
\name{deleteLibrary}
\alias{deleteLibrary}
\title{Delete Library}
\usage{
deleteLibrary(
  ssimLibrary,
  force = FALSE,
  removeBackup = FALSE,
  removePublish = FALSE,
  removeCustom = FALSE,
  session = NULL
)
}
\arguments{
\item{ssimLibrary}{SsimLibrary or path to a library}

\item{force}{Logical. If FALSE (default) prompt to confirm that the library
should be deleted. This is irreversable.}

\item{removeBackup}{logical. If \code{TRUE}, will remove the backup folder when
deleting a library. Default is FALSE.}

\item{removePublish}{logical. If TRUE, will remove the publish folder when
deleting a library. Default is FALSE.}

\item{removeCustom}{logical. If TRUE and custom folders have been configured
for a library, then will remove the custom publish and/or backup folders when
deleting a library. Note that the \code{removePublish} and \code{removeBackup} arguments
must also be set to TRUE to remove the respective custom folders. Default
is FALSE.}

\item{session}{Session}
}
\value{
"saved" or failure message.
}
\description{
Deletes a SyncroSim library. Note this is irreversable.
}
