% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimLibrary.R
\name{deleteLibrary}
\alias{deleteLibrary}
\title{Delete Library}
\usage{
deleteLibrary(ssimLibrary, force = F)
}
\arguments{
\item{ssimLibrary}{SsimLibrary or path to a library}

\item{force}{Logical. If FALSE (default) prompt to confirm that the library should be deleted. This is irreversable.}
}
\value{
"saved" or failure message.
}
\description{
Deletes a SyncroSim library. Note this is irreversable.
}
\examples{

}

