% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalHelpers.R
\name{deleteLibrary}
\alias{deleteLibrary}
\title{deleting a library. Default is FALSE
deleting a library. Default is FALSE
for a library, then will remove the custom publish and/or backup folders when
deleting a library. Note that the \code{removePublish} and \code{removeBackup} arguments
must also be set to TRUE to remove the respective custom folders. Default
is FALSE}
\usage{
deleteLibrary(
  ssimLibrary,
  force = FALSE,
  removeBackup = FALSE,
  removePublish = FALSE,
  removeCustom = FALSE,
  session = NULL
)
}
\description{
deleting a library. Default is FALSE
deleting a library. Default is FALSE
for a library, then will remove the custom publish and/or backup folders when
deleting a library. Note that the \code{removePublish} and \code{removeBackup} arguments
must also be set to TRUE to remove the respective custom folders. Default
is FALSE
}
