% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteModule.R
\name{deleteModule}
\alias{deleteModule}
\alias{deleteModule,ANY,missingOrNULLOrChar-method}
\alias{deleteModule,ANY,Session-method}
\title{Delete module or modules}
\usage{
deleteModule(name, session = NULL, force = F)

\S4method{deleteModule}{ANY,missingOrNULLOrChar}(name, session = NULL, force = F)

\S4method{deleteModule}{ANY,Session}(name, session = NULL, force = F)
}
\arguments{
\item{name}{Character string or vector of these. A module or vector of modules to remove. See modules() for options.}

\item{session}{Session.}

\item{force}{logical. If T, delete without requiring confirmation from user.}
}
\value{
"saved" or error message.
}
\description{
Deprecated.  See: \code{\link{deletePackage}}
}
