% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deletePackage}
\alias{deletePackage}
\title{Deletes a package from your SyncroSim installation}
\usage{
deletePackage(name, session = NULL, force = FALSE)
}
\arguments{
\item{name}{character string or vector of these. A package or vector of
packages to remove}

\item{session}{\code{\link{Session}} object}

\item{force}{logical. If \code{FALSE} (default), require confirmation from user
before deletion}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Please use \code{\link{removePackage}} instead.
}
\keyword{internal}
