% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletePackage.R
\name{deletePackage}
\alias{deletePackage}
\alias{deletePackage,ANY,character-method}
\alias{deletePackage,ANY,missingOrNULL-method}
\alias{deletePackage,ANY,Session-method}
\title{Deletes a package}
\usage{
deletePackage(name, session = NULL, force = F)

\S4method{deletePackage}{ANY,character}(name, session = NULL, force = F)

\S4method{deletePackage}{ANY,missingOrNULL}(name, session = NULL, force = F)

\S4method{deletePackage}{ANY,Session}(name, session = NULL, force = F)
}
\arguments{
\item{name}{Character. The name of the package to delete.}

\item{session}{Session.}

\item{force}{logical. If T, delete without requiring confirmation from user.}
}
\value{
"saved" or error message.
}
\description{
Deletes a package.
}
