% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deletePackage}
\alias{deletePackage}
\title{Deletes a package from your SyncroSim installation}
\usage{
deletePackage(name, session = NULL, force = FALSE)
}
\arguments{
\item{name}{character string or vector of these. A package or vector of 
packages to remove}

\item{session}{\code{\link{Session}} object}

\item{force}{logical. If \code{FALSE} (default), require confirmation from user
before deletion}
}
\description{
`r lifecycle::badge("deprecated")`
Please use \code{\link{removePackage}} instead.
}
\keyword{internal}
