% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimLibrary.R
\name{deleteProjects}
\alias{deleteProjects}
\title{Delete projects from a Library}
\usage{
deleteProjects(x, project = NULL, force = F)
}
\arguments{
\item{x}{An SsimLibrary, Project or Scenario associated with a library.}

\item{project}{One or more project names or ids.}
}
\value{
A list of "saved" or failure messages for each project.
}
\description{
Deletes one or more projects from a SyncroSim library.
}
\examples{
myLibrary = ssimLibrary(session=devSession)
myProject = project(myLibrary)
projects(myLibrary,names=T)
deleteProjects(myLibrary,project="Project1")
projects(myLibrary,names=T)

}

