% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency.R
\docType{methods}
\name{dependency}
\alias{dependency}
\alias{dependency,character-method}
\alias{dependency,Scenario-method}
\title{Set or remove Scenario dependency(s), or get existing dependencies.}
\usage{
dependency(scenario, dependency = NULL, remove = F, force = F)

\S4method{dependency}{character}(scenario, dependency = NULL, remove = F,
  force = F)

\S4method{dependency}{Scenario}(scenario, dependency = NULL, remove = F,
  force = F)
}
\arguments{
\item{scenario}{Scenario. The scenario to which a dependency is to be added (or has already been added if remove=TRUE).}

\item{dependency}{Scenario, character string, integer, or list/vector of these. The scenario(s) that are the source of the dependency, in order from lowest to highest precedence. If NULL other arguments are ingored and the list of existing dependencies is returned.}

\item{remove}{logical. If F (default) dependencies are added. If T, dependencies are removed.}

\item{force}{logical. If F (default) prompt before removing dependencies.}
}
\value{
If dependency!=NULL, character string (saved or error message) or list of these. Otherwise, a dataframe of existing dependencies, or list of these.
}
\description{
Set or remove Scenario dependency(s), or get existing dependencies.
}
\details{
If dependency==NULL, other arguments are ignored, and set of existing dependencies is returned in order of precedence (from highest to lowest precedence).
Otherwise, returns list of saved or error messages for each dependency of each scenario.

Note that the order of dependencies can be important - dependencies added most recently take precedence over existing dependencies.
So, dependencies included in the dependency argument take precedence over any other existing dependencies.
If the dependency argument includes more than one element, elements are ordered from lowest to highest precedence.
}
