% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{description<-}
\alias{description<-}
\title{Set the description of an SsimLibrary/Project/Scenario.}
\usage{
description(ssimObject) <- value
}
\arguments{
\item{ssimObject}{Scenario/Project/SsimLibrary}

\item{value}{The new description.}
}
\description{
Set the description of an SsimLibrary/ProjectScenario.
}

