% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/scenario.R
\name{description}
\alias{description}
\title{Description of an SsimLibrary/Project/Scenario.}
\usage{
description(x)

description(x)
}
\arguments{
\item{x}{An Scenario object.}

\item{ssimObject}{SsimLibrary/Project/Scenario.}
}
\value{
The description.
}
\description{
The description of an SsimLibrary/ProjectScenario.

The description of a Scenario
}

