% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{description}
\alias{description}
\title{Description of an SsimLibrary/Project/Scenario.}
\usage{
description(ssimObject)
}
\arguments{
\item{ssimObject}{SsimLibrary/Project/Scenario.}
}
\description{
The description of an SsimLibrary/ProjectScenario.
}

