% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{description}
\alias{description}
\alias{description,character-method}
\alias{description,SsimObject-method}
\title{Description of an SsimLibrary/Project/Scenario.}
\usage{
description(ssimObject)

\S4method{description}{character}(ssimObject)

\S4method{description}{SsimObject}(ssimObject)
}
\arguments{
\item{ssimObject}{SsimLibrary/Project/Scenario.}
}
\description{
The description of an SsimLibrary/ProjectScenario.
}
