% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{description}
\alias{description}
\title{The description of a Scenario}
\usage{
description(x)
}
\arguments{
\item{x}{An Scenario object.}
}
\value{
The description.
}
\description{
The description of a Scenario
}

