% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{description}
\alias{description}
\alias{description,character-method}
\alias{description,SsimObject-method}
\title{Description of a SsimLibrary, Project or Scenario.}
\usage{
description(ssimObject)

\S4method{description}{character}(ssimObject)

\S4method{description}{SsimObject}(ssimObject)
}
\arguments{
\item{ssimObject}{An object of class SsimLibrary, Project or Scenario.}
}
\value{
A character string describing the ssimObject.
}
\description{
Returns a description of a \code{\link{SsimLibrary}}, \code{\link{Project}} or 
\code{\link{Scenario}}.
}
\examples{
\donttest{
temp_dir <- tempdir()
myses <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"), session = myses)
myProject <- project(myLibrary)

mydescription <- description(myProject)
}

}
