% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\docType{methods}
\name{description}
\alias{description}
\alias{description,Project-method}
\alias{description,Scenario-method}
\alias{description,SsimLibrary-method}
\title{Description of an SsimLibrary/Project/Scenario.}
\usage{
description(ssimObject)

\S4method{description}{SsimLibrary}(ssimObject)

\S4method{description}{Project}(ssimObject)

\S4method{description}{Scenario}(ssimObject)
}
\arguments{
\item{ssimObject}{SsimLibrary/Project/Scenario.}
}
\description{
The description of an SsimLibrary/ProjectScenario.
}
\section{Methods (by class)}{
\itemize{
\item \code{SsimLibrary}: description of an SsimLibrary.

\item \code{Project}: description of a Project.

\item \code{Scenario}: description of a Scenrio.
}}

