% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disableAddon.R
\name{disableAddon}
\alias{disableAddon}
\alias{disableAddon,character-method}
\alias{disableAddon,SsimLibrary-method}
\title{Disable addon or addons.}
\usage{
disableAddon(ssimLibrary, name)

\S4method{disableAddon}{character}(ssimLibrary, name)

\S4method{disableAddon}{SsimLibrary}(ssimLibrary, name)
}
\arguments{
\item{ssimLibrary}{An object of class SsimLibrary.}

\item{name}{Character string or vector of addon names.}
}
\value{
This function invisibly returns \code{TRUE} upon success (i.e.successful deactivation
of the addon) or \code{FALSE} upon failure.
}
\description{
Disable \code{\link{addon}} or addons of a \code{\link{SsimLibrary}},
\code{\link{Project}} or \code{\link{Scenario}} with an associated SsimLibrary.
}
\examples{
\donttest{
addPackage("stsim")
temp_dir <- tempdir()
mySession <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"),
                         session = mySession, package = "stsim")

# Enable add on package
enableAddon(myLibrary, c("stsimsf"))
addon(myLibrary)

# Disable add on package
disableAddon(myLibrary, c("stsimsf"))
addon(myLibrary)
}

}
\seealso{
\code{\link{addon}}
}
