% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disableAddon.R
\docType{methods}
\name{disableAddon}
\alias{disableAddon}
\alias{disableAddon,SsimLibrary-method}
\title{Disable addon or addons.}
\usage{
disableAddon(ssimLibrary, name)

\S4method{disableAddon}{SsimLibrary}(ssimLibrary, name)
}
\arguments{
\item{ssimLibrary}{SsimLibrary}

\item{name}{Character string or vector of these.}
}
\value{
saved or error message.
}
\description{
Disable addon or addons of an SsimLibrary, or Project/Scenario with an associated SsimLibrary.
}
\examples{
#TODO - update examples
myLibrary = ssimLibrary()
enableAddon(myLibrary,c("stsim-ecological-departure"))
addon(myLibrary)
disableAddon(myLibrary,c("stsim-ecological-departure"))
addon(myLibrary)

}

