% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disableAddon.R
\name{disableAddon}
\alias{disableAddon}
\alias{disableAddon,character-method}
\alias{disableAddon,SsimLibrary-method}
\title{Disable addon or addons.}
\usage{
disableAddon(ssimLibrary, name)

\S4method{disableAddon}{character}(ssimLibrary, name)

\S4method{disableAddon}{SsimLibrary}(ssimLibrary, name)
}
\arguments{
\item{ssimLibrary}{SsimLibrary}

\item{name}{Character string or vector of addon names}
}
\value{
Character "saved" in case of success or error message.
}
\description{
Disable addon or addons of an SsimLibrary, or Project/Scenario with an associated SsimLibrary.
}
\examples{
myLibrary <- ssimLibrary("mylib")
enableAddon(myLibrary, c("stsimecodep"))
addon(myLibrary)
\dontrun{
disableAddon(myLibrary, c("stsimecodep"))
}
addon(myLibrary)
}
