% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enableAddon.R
\name{enableAddon}
\alias{enableAddon}
\alias{enableAddon,character-method}
\alias{enableAddon,SsimLibrary-method}
\title{Enable addon or addons.}
\usage{
enableAddon(ssimLibrary, name)

\S4method{enableAddon}{character}(ssimLibrary, name)

\S4method{enableAddon}{SsimLibrary}(ssimLibrary, name)
}
\arguments{
\item{ssimLibrary}{An object of class SsimLibrary.}

\item{name}{Character string or vector of addon names.}
}
\value{
This function invisibly returns \code{TRUE} upon success (i.e.successful activation
of the addon) or \code{FALSE} upon failure.
}
\description{
Enable \code{\link{addon}} or addons of a \code{\link{SsimLibrary}}.
}
\examples{
\donttest{
addPackage("stsim")
temp_dir <- tempdir()
mySession <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"),
                         session = mySession, package = "stsim")

# Enable add on package
enableAddon(myLibrary, c("stsimsf"))
addon(myLibrary)

# Disable add on package
disableAddon(myLibrary, c("stsimsf"))
addon(myLibrary)
}

}
\seealso{
\code{\link{addon}}
}
