% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enableAddon.R
\name{enableAddon}
\alias{enableAddon}
\alias{enableAddon,character-method}
\alias{enableAddon,SsimLibrary-method}
\title{Enable addon or addons.}
\usage{
enableAddon(ssimLibrary, name)

\S4method{enableAddon}{character}(ssimLibrary, name)

\S4method{enableAddon}{SsimLibrary}(ssimLibrary, name)
}
\arguments{
\item{ssimLibrary}{SsimLibrary}

\item{name}{Character string or vector of these.}
}
\value{
saved or error message for each addon.
}
\description{
Enable addon or addons of an SsimLibrary.
}
\examples{
myLibrary <- ssimLibrary()
enableAddon(myLibrary, c("stsim-ecological-departure"))
addon(myLibrary)
\dontrun{
disableAddon(myLibrary, c("stsim-ecological-departure"))
}
addon(myLibrary)
}
