% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{envBeginSimulation}
\alias{envBeginSimulation}
\title{Begins a SyncroSim simulation}
\usage{
envBeginSimulation(totalSteps)
}
\arguments{
\item{totalSteps}{integer.  The total number of steps in the simulation.}
}
\description{
Begins a SyncroSim simulation
}
