% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{envBeginSimulation}
\alias{envBeginSimulation}
\title{Begins a SyncroSim simulation}
\usage{
envBeginSimulation(totalSteps)
}
\arguments{
\item{totalSteps}{integer.  The total number of steps in the simulation}
}
\value{
No returned value, used for side effects.
}
\description{
This function is part of a set of functions designed to facilitate the
development of R-based SyncroSim Packages. Specifically, this function begins
the reporting of a SyncroSim simulation.
}
\examples{
\dontrun{
n_iter <- 50
n_timesteps <- 10
n_steps <- n_iter * n_timesteps
envBeginSimulation(n_steps)
}

}
