% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{envEndSimulation}
\alias{envEndSimulation}
\title{Ends a SyncroSim simulation}
\usage{
envEndSimulation()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Please use \code{\link{progressBar}} instead.
}
\keyword{internal}
