% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{envEndSimulation}
\alias{envEndSimulation}
\title{Ends a SyncroSim simulation}
\usage{
envEndSimulation()
}
\value{
No returned value, used for side effects.
}
\description{
This function is part of a set of functions designed to facilitate the
development of R-based Syncrosim Packages. This function ends a SyncroSim 
simulation.
}
\examples{
\dontrun{
envEndSimulation()
}

}
