% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{envInputFolder}
\alias{envInputFolder}
\title{SyncroSim DataSheet Input Folder}
\usage{
envInputFolder(scenario, datasheetName)
}
\arguments{
\item{scenario}{\code{\link{Scenario}} object. A SyncroSim result Scenario}

\item{datasheetName}{character. The input Datasheet name}
}
\value{
Returns a folder name for the specified Datasheet.
}
\description{
This function is part of a set of functions designed to facilitate the
development of R-based Syncrosim Packages. This function creates and returns 
a SyncroSim Datasheet Input Folder.
}
\examples{
\dontrun{
inputFolder <- envInputFolder
}

}
