% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{envInputFolder}
\alias{envInputFolder}
\title{SyncroSim DataSheet Input Folder}
\usage{
envInputFolder(scenario, datasheetName)
}
\arguments{
\item{scenario}{Scenario.  A SyncroSim result scenario.}

\item{datasheetName}{character.  The input datasheet name.}
}
\value{
a folder name for the specified data sheet
}
\description{
Creates and returns a SyncroSim DataSheet Input Folder
}
