% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{envInputFolder}
\alias{envInputFolder}
\title{SyncroSim DataSheet Input Folder}
\usage{
envInputFolder(scenario, datasheetName)
}
\arguments{
\item{scenario}{Scenario.  A SyncroSim result \code{\link{Scenario}}.}

\item{datasheetName}{Character.  The input datasheet name.}
}
\value{
Returns a folder name for the specified datasheet.
}
\description{
This function is part of a set of functions designed to facilitate the
development of R-based Syncrosim Packages. This function creates and returns 
a SyncroSim Datasheet Input Folder.
}
\examples{
\dontrun{
inputFolder <- envInputFolder
}

}
