% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{envOutputFolder}
\alias{envOutputFolder}
\title{SyncroSim DataSheet Output Folder}
\usage{
envOutputFolder(scenario, datasheetName)
}
\arguments{
\item{scenario}{Scenario. A SyncroSim result \code{\link{Scenario}}.}

\item{datasheetName}{Character. The output datasheet name.}
}
\value{
Returns a folder name for the specified datasheet.
}
\description{
This function is part of a set of functions designed to facilitate the
development of R-based Syncrosim Packages. Thus function creates and returns 
a SyncroSim DataSheet Output Folder.
}
\examples{
\donttest{
outputFolder <- envOutputFolder()
}

}
