% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{envReportProgress}
\alias{envReportProgress}
\title{Reports SyncroSim simulation progress}
\usage{
envReportProgress(iteration, timestep)
}
\arguments{
\item{iteration}{integer. The current iteration}

\item{timestep}{integer. The current timestep}
}
\value{
No returned value, used for side effects.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
Please use \code{\link{progressBar}} instead.
}
\keyword{internal}
