% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{envReportProgress}
\alias{envReportProgress}
\title{Reports SyncroSim simulation progress}
\usage{
envReportProgress(iteration, timestep)
}
\arguments{
\item{iteration}{integer. The current iteration}

\item{timestep}{integer. The current timestep}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Please use \code{\link{progressBar}} instead.
}
\keyword{internal}
