% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{envReportProgress}
\alias{envReportProgress}
\title{Reports progress for a SyncroSim simulation}
\usage{
envReportProgress(iteration, timestep)
}
\arguments{
\item{iteration}{integer. The current iteration.}

\item{timestep}{integer. The current timestep.}
}
\value{
No returned value, used for side effects.
}
\description{
This function is part of a set of functions designed to facilitate the
development of R-based Syncrosim Packages. This function reports progress 
for a SyncroSim simulation.
}
\examples{
\dontrun{
envReportProgress()
}

}
