% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{envStepSimulation}
\alias{envStepSimulation}
\title{Steps a SyncroSim simulation}
\usage{
envStepSimulation()
}
\description{
`r lifecycle::badge("deprecated")`
Please use \code{\link{progressBar}} instead.
}
\keyword{internal}
