% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{envTempFolder}
\alias{envTempFolder}
\title{SyncroSim Temporary Folder}
\usage{
envTempFolder(folderName)
}
\arguments{
\item{folderName}{character. The folder name}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Please use \code{\link{runtimeTempFolder}} instead.
}
\keyword{internal}
