% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{envTempFolder}
\alias{envTempFolder}
\title{SyncroSim Temporary Folder}
\usage{
envTempFolder(folderName)
}
\arguments{
\item{folderName}{Character. The folder name.}
}
\value{
Returns a temporary folder name.
}
\description{
This function is part of a set of functions designed to facilitate the
development of R-based Syncrosim Packages. This function creates and returns
a SyncroSim Temporary Folder.
}
\examples{
\dontrun{
tempFolder <- envTempFolder()
}

}
