% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/session.R, R/ssimLibrary.R
\docType{methods}
\name{filepath}
\alias{filepath}
\alias{filepath,SSimLibrary-method}
\alias{filepath,Session-method}
\title{The path to an object on disk}
\usage{
filepath(x)

\S4method{filepath}{Session}(x)

\S4method{filepath}{SSimLibrary}(x)
}
\arguments{
\item{x}{An object containing a path to a file on disk.}
}
\description{
The path to a SyncroSim Session, SSimLibarary, Project or Scenario on disk.
}
\section{Methods (by class)}{
\itemize{
\item \code{Session}: Path to the Syncrosim console in a Session.

\item \code{SSimLibrary}: Path to an SSimLibrary on disk.
}}

