% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath.R
\name{filepath}
\alias{filepath}
\title{The path to a SyncroSim object on disk}
\usage{
filepath(ssimObject)
}
\arguments{
\item{ssimObject}{An object containing a filepath.}
}
\description{
The path to a SyncroSim Session, SSimLibarary, Project or Scenario on disk.
}

