% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R, R/folder.R
\name{folder}
\alias{folder}
\alias{chart}
\title{Create or open a chart}
\usage{
chart(ssimObject = NULL, chart = NULL, create = FALSE, summary = FALSE)

folder(ssimObject = NULL, folder = NULL, parentFolder = NULL, create = FALSE)
}
\arguments{
\item{ssimObject}{\code{\link{SsimLibrary}} or \code{\link{Project}} object.}

\item{chart}{character or integer. If character, then will either open an
existing chart if \code{create=FALSE}, or will create a new chart with the
given name if the chart does not exist yet or \code{create=TRUE}.
If integer, will open the existing chart with the given chart ID (if the
ID exists). If no value is provided and \code{create=TRUE}, a new chart will
be created with the default naming convention (e.g. "_Chart1", "_Chart2")}

\item{create}{logical. Whether to create a new folder if the folder name given
already exists in the SyncroSim library. If \code{FALSE} (Default), then will
return the existing folder with the given name. If \code{TRUE}, then will
return a new folder with the same name as an existing folder (but different
folder ID)}

\item{summary}{logical. If \code{TRUE}, returns a summary of chart
information as an R data.frame. If \code{FALSE} (Default), then returns
a SyncroSim Chart object}

\item{folder}{character or integer. If character, then will either open an
existing folder if \code{create=FALSE}, or will create a new folder with the
given name if the folder does not exist yet or \code{create=TRUE} (Default).
If integer, will open the existing folder with the given folder ID (if the
ID exists).}

\item{parentFolder}{character, integer, or SyncroSim Folder object. If not
\code{NULL} (Default), the new folder will be created inside of the
specified parent folder}
}
\value{
A \code{Chart} object representing a SyncroSim chart

A \code{Folder} object representing a SyncroSim folder.
}
\description{
Create or open a \code{\link{Chart}} from a SyncroSim
\code{\link{Project}}.

Create or open a \code{\link{Folder}} from a SyncroSim
\code{\link{Project}}.
}
\examples{
\donttest{
# Set the file path and name of the new SsimLibrary
myLibraryName <- file.path(tempdir(),"testlib")

# Set the SyncroSim Session, SsimLibrary, Project, and Scenario
mySession <- session()
myLibrary <- ssimLibrary(name = myLibraryName, session = mySession) 
myProject <- project(myLibrary, project = "My Project")
myScenario <- scenario(myProject, scenario = "My Scenario")

# Create a new chart
myChart <- chart(myProject, chart = "New Chart")
}
\donttest{
# Set the file path and name of the new SsimLibrary
myLibraryName <- file.path(tempdir(),"testlib")

# Set the SyncroSim Session, SsimLibrary, Project, and Scenario
mySession <- session()
myLibrary <- ssimLibrary(name = myLibraryName, session = mySession) 
myProject <- project(myLibrary, project = "My Project")
myScenario <- scenario(myProject, scenario = "My Scenario")

# Create a new folder
myFolder <- folder(myProject, folder = "New Folder")

# Create a nested folder within "New Folder"
myNestedFolder <- folder(myProject, folder = "New Nested Folder", 
                         parentFolder = myFolder)
}

}
