% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignoreDependencies.R
\name{ignoreDependencies}
\alias{ignoreDependencies}
\alias{ignoreDependencies,character-method}
\alias{ignoreDependencies,Scenario-method}
\title{Ignore Dependencies for a Scenario.}
\usage{
ignoreDependencies(ssimObject)

\S4method{ignoreDependencies}{character}(ssimObject)

\S4method{ignoreDependencies}{Scenario}(ssimObject)
}
\arguments{
\item{ssimObject}{An object of class \code{\link{Scenario}}.}
}
\value{
Returns Ignore Dependencies Datafeeds.
}
\description{
Retrieves the Ignore Dependencies Datafeeds for a \code{\link{Scenario}}.
}
\examples{
\donttest{
temp_dir <- tempdir()
mySession <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"), session = mySession)
myProject <- project(myLibrary, project = "Definitions")
myScenario <- scenario(myProject, scenario = "My Scenario")

ignoreDependencies(myScenario)
}

}
