% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignoreDependencies.R
\name{ignoreDependencies}
\alias{ignoreDependencies}
\alias{ignoreDependencies,character-method}
\alias{ignoreDependencies,Scenario-method}
\title{Ignore Dependencies for a Scenario.}
\usage{
ignoreDependencies(ssimObject)

\S4method{ignoreDependencies}{character}(ssimObject)

\S4method{ignoreDependencies}{Scenario}(ssimObject)
}
\arguments{
\item{ssimObject}{Scenario.}
}
\value{
Returns Ignore Dependencies Datafeeds.
}
\description{
Retrieves the Ignore Dependencies Datafeeds for a \code{\link{Scenario}}.
}
