% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/ssimLibrary.R
\docType{methods}
\name{info}
\alias{info}
\alias{info,SSimLibrary-method}
\title{Information about an object}
\usage{
info(x)

\S4method{info}{SSimLibrary}(x)
}
\arguments{
\item{x}{An object containing info.}
}
\description{
Get basic information about a SyncroSim Session, SSimLibarary, Project or Scenario
}
\section{Methods (by class)}{
\itemize{
\item \code{SSimLibrary}: Info about an SSimLibrary.
}}

