% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\alias{info,SsimLibrary-method}
\title{Information about an library}
\usage{
info(ssimLibrary)

\S4method{info}{SsimLibrary}(ssimLibrary)
}
\arguments{
\item{ssimLibrary}{A \code{SsimLibrary} object.}
}
\value{
Returns a \code{data.frame} with information on the properties of the library object.
}
\description{
Get basic information about a Library.
}
