% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{info}
\alias{info}
\title{Information about an object}
\usage{
info(x)
}
\arguments{
\item{x}{An object containing info.}
}
\description{
Get basic information about a SyncroSim Session, SSimLibarary, Project or Scenario
}

