% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\alias{info,SsimLibrary-method}
\title{Retrieves information about an library}
\usage{
info(ssimLibrary)

\S4method{info}{SsimLibrary}(ssimLibrary)
}
\arguments{
\item{ssimLibrary}{An object of type \code{\link{SsimLibrary}}.}
}
\value{
Returns a \code{data.frame} with information on the properties of the library 
object.
}
\description{
Retrieves some basic metadata about a Library.
}
\examples{
\donttest{
temp_dir <- tempdir()
mySession <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"), session = mySession)

info(myLibrary)
}

}
