% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\alias{info,SsimLibrary-method}
\title{Retrieves information about a library}
\usage{
info(ssimLibrary)

\S4method{info}{SsimLibrary}(ssimLibrary)
}
\arguments{
\item{ssimLibrary}{An object of class \code{\link{SsimLibrary}}.}
}
\value{
Returns a \code{data.frame} with information on the properties of the library 
object.
}
\description{
Retrieves some basic metadata about a Library: Name, Owner, Last Modified, 
Size, Read Only, Package Name, Package Description, Current Package Version,
Minimum Package Version, External input files, External output files, 
Temporary files, Backup files.
}
\examples{
\dontrun{
temp_dir <- tempdir()
mySession <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"), session = mySession)

info(myLibrary)
}

}
