% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installConda.R
\name{installConda}
\alias{installConda}
\alias{installConda,character-method}
\alias{installConda,missingOrNULL-method}
\alias{installConda,Session-method}
\title{Installs Miniconda}
\usage{
installConda(session = NULL)

\S4method{installConda}{character}(session = NULL)

\S4method{installConda}{missingOrNULL}(session = NULL)

\S4method{installConda}{Session}(session = NULL)
}
\arguments{
\item{session}{\code{\link{Session}} object. If \code{NULL} (default),
\code{session()} will be used}
}
\value{
Invisibly returns \code{TRUE} upon success (i.e.successful 
install) and \code{FALSE} upon failure.
}
\description{
This function installs Miniconda to the default installation path
within the SyncroSim installation folder.
}
\examples{
\dontrun{
# Create a new SyncroSim Session
mySession <- session()

# Install Conda for the given SyncroSim session
installConda(mySession)
}

}
