% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installConda.R
\name{installConda}
\alias{installConda}
\title{Installs Miniconda}
\usage{
installConda()
}
\arguments{
\item{session}{\code{\link{Session}} object. If \code{NULL} (default),
\code{session()} will be used}
}
\value{
Invisibly returns \code{TRUE} upon success (i.e.successful 
install) and \code{FALSE} upon failure.
}
\description{
This function installs Miniconda to the default installation path
within the SyncroSim installation folder.
}
\examples{
\dontrun{
# Install Conda for the default SyncroSim session
installConda()
}

}
