% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installConda.R
\name{installConda}
\alias{installConda}
\alias{installConda,ANY,character-method}
\alias{installConda,ANY,missingOrNULL-method}
\alias{installConda,ANY,Session-method}
\title{Installs Miniconda}
\usage{
installConda(folderPath, session = NULL)

\S4method{installConda}{ANY,character}(folderPath, session = NULL)

\S4method{installConda}{ANY,missingOrNULL}(folderPath, session = NULL)

\S4method{installConda}{ANY,Session}(folderPath, session = NULL)
}
\arguments{
\item{folderPath}{character string.  The file path to the Conda installation folder.
If \code{NULL}, then the default installation folder is used}

\item{session}{\code{\link{Session}} object. If \code{NULL} (default),
\code{session()} will be used}
}
\value{
Invisibly returns \code{TRUE} upon success (i.e.successful 
install) and \code{FALSE} upon failure.
}
\description{
This function installs Miniconda to either the default installation path
within the SyncroSim installation folder or a custom path.
}
\examples{
\dontrun{
# Create a new SyncroSim Session
mySession <- session()

# Install Conda for the given SyncroSim session
installConda(folderPath = "C:/miniconda3", mySession)
}

}
