% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{loadDatasheets}
\alias{loadDatasheets}
\title{Set datasheets}
\usage{
loadDatasheets(x, ...)
}
\arguments{
\item{x}{An SSimLibrary, Project or Scenario object. Or the path to a library on disk.}

\item{data}{A dataframe or named list of dataframes to load.}

\item{name}{The sheet name - required if data is a dataframe, ignored otherwise.}

\item{project}{Project name or id.}

\item{scenario}{Scenario name or id.}
}
\value{
A named list of success or failure reports.
}
\description{
Loads datasheets into the SyncroSim library.
}
\examples{

}

