% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{loadSpatialData}
\alias{loadSpatialData}
\title{Set spatial data}
\usage{
loadSpatialData(x, data, metadata = NULL, project = NULL, scenario = NULL,
  breakpoint = F)
}
\arguments{
\item{x}{An SSimLibrary, Project or Scenario object. Or the path to a library on disk.}

\item{data}{A RasterLayer or RasterStack to load.}

\item{metadata}{A dataframe that can be appended to datasheet(x,metadata$SheetName[1]), containing 1 row for each layer - see details. If NULL, use names(data) metadata.}

\item{project}{Project name or id.}

\item{scenario}{Scenario name or id.}

\item{breakpoint}{Set to TRUE when setting spatial data in a breakpoint function.}
}
\value{
A named list of success or failure reports.
}
\description{
Loads spatial data into the SyncroSim library.
}
\details{
If metadata=NULL or sheetName=NULL assume each raster layer has names() metadata - see spatialData() for details.

Otherwise, metadata should be a dataframe that can be appended to datasheet(x,metadata$SheetName[1]), containing 1 row for each layer of data.
"SheetName" and "RasterLayerName" columns are expected in metadata, and will be removed before appending to the datasheet.

INCOMPLETE: this method has only been implemented for breakpoint=T and sheet name is STSim_TransitionSpatialMultiplier
}
\examples{

}

