% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeDependencies.R
\name{mergeDependencies<-}
\alias{mergeDependencies<-}
\alias{mergeDependencies<-,character-method}
\alias{mergeDependencies<-,Scenario-method}
\title{Merge Dependencies for a Scenario.}
\usage{
mergeDependencies(ssimObject) <- value

\S4method{mergeDependencies}{character}(ssimObject) <- value

\S4method{mergeDependencies}{Scenario}(ssimObject) <- value
}
\arguments{
\item{ssimObject}{Scenario.}

\item{value}{Logical. If TRUE the Scenario will be set to merge dependencies at run time.}
}
\value{
Returns the updated ssimObject.
}
\description{
Sets whether or not a Scenario is configured to merge dependencies at run time.
}
