% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeDependencies.R
\name{mergeDependencies}
\alias{mergeDependencies}
\alias{mergeDependencies,character-method}
\alias{mergeDependencies,Scenario-method}
\title{Merge Dependencies for a Scenario.}
\usage{
mergeDependencies(ssimObject)

\S4method{mergeDependencies}{character}(ssimObject)

\S4method{mergeDependencies}{Scenario}(ssimObject)
}
\arguments{
\item{ssimObject}{Scenario}
}
\value{
Returns a logical: `TRUE` is the scenario is configured to merge dependencies at run time, 
and `FALSE` otherwise.
}
\description{
Retrieves whether or not a Scenario is configured to merge dependencies at run time.
}
