% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{model}
\alias{model}
\title{Installed models}
\usage{
model(ssimObject = NULL)
}
\arguments{
\item{ssimObject}{\code{\link{Session}} or \code{\link{SsimLibrary}} object}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Models are now distributed in Packages;
please use \code{\link{package}} instead.
}
\keyword{internal}
