% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Installed models}
\usage{
model(ssimObject = NULL)
}
\arguments{
\item{ssimObject}{Session or SsimLibrary.}
}
\value{
A dataframe of models (for Session) or named vector of character strings (for SsimLibrary)
}
\description{
Models installed with this version of SyncroSim
}

