% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{model}
\alias{model}
\title{Installed models}
\usage{
model(ssimObject = NULL)
}
\arguments{
\item{ssimObject}{\code{\link{Session}} or \code{\link{SsimLibrary}}.}
}
\value{
A \code{dataframe} of models (for Session) or named vector of character strings
(for \code{SsimLibrary}).
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
models are now distributed in Packages;
Please use \code{\link{package}} instead.
}
