% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{methods}
\name{model}
\alias{model}
\alias{model,Session-method}
\alias{model,SsimLibrary-method}
\alias{model,missingOrNULL-method}
\title{Installed models}
\usage{
model(ssimObject = NULL)

\S4method{model}{missingOrNULL}(ssimObject = NULL)

\S4method{model}{Session}(ssimObject = NULL)

\S4method{model}{SsimLibrary}(ssimObject = NULL)
}
\arguments{
\item{ssimObject}{Session or SsimLibrary.}
}
\value{
A dataframe of models (for Session) or named vector of character strings (for SsimLibrary)
}
\description{
Models installed with this version of SyncroSim
}

