% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\alias{model,character-method}
\alias{model,missingOrNULL-method}
\alias{model,Session-method}
\alias{model,SsimLibrary-method}
\title{Installed models}
\usage{
model(ssimObject = NULL)

\S4method{model}{character}(ssimObject = NULL)

\S4method{model}{missingOrNULL}(ssimObject = NULL)

\S4method{model}{Session}(ssimObject = NULL)

\S4method{model}{SsimLibrary}(ssimObject = NULL)
}
\arguments{
\item{ssimObject}{\code{\link{Session}} or \code{\link{SsimLibrary}}.}
}
\value{
A \code{dataframe} of models (for Session) or named vector of character strings 
(for \code{SsimLibrary}).
}
\description{
Deprecated.  See: \code{\link{package}}.
}
