% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimLibrary.R
\name{modelName}
\alias{modelName}
\title{The name of the primary model associate with a SyncroSim object}
\usage{
modelName(x)
}
\arguments{
\item{x}{An object with an associated primary model.}
}
\value{
A model name
}
\description{
The name of the primary model associated with a SSimLibarary, Project or Scenario.
}

