% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{module}
\alias{module}
\title{Installed modules}
\usage{
module(session = NULL)
}
\arguments{
\item{session}{\code{\link{Session}} object}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
modules are now distributed in Packages; 
Please use \code{\link{package}} instead.
}
\keyword{internal}
