% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{module}
\alias{module}
\alias{module,missingOrNULL-method}
\alias{module,character-method}
\alias{module,Session-method}
\title{Installed modules}
\usage{
module(session)

\S4method{module}{missingOrNULL}(session)

\S4method{module}{character}(session)

\S4method{module}{Session}(session)
}
\arguments{
\item{session}{Session.}
}
\value{
A \code{dataframe} of modules.
}
\description{
Deprecated.  See: \code{\link{package}}
}
