% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\docType{methods}
\name{module}
\alias{module}
\alias{module,Session-method}
\alias{module,character-method}
\alias{module,missingOrNULL-method}
\title{Installed modules}
\usage{
module(session)

\S4method{module}{missingOrNULL}(session)

\S4method{module}{character}(session)

\S4method{module}{Session}(session)
}
\arguments{
\item{session}{Session.}
}
\value{
A dataframe of modules
}
\description{
Modules installed with this version of SyncroSim
}

