% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{methods}
\name{modules}
\alias{modules}
\alias{modules,Session-method}
\alias{modules,missingOrNULL-method}
\title{Installed modules}
\usage{
modules(session)

\S4method{modules}{missingOrNULL}(session)

\S4method{modules}{Session}(session)
}
\arguments{
\item{session}{Session.}
}
\value{
A dataframe of modules
}
\description{
Modules installed with this version of SyncroSim
}

