% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{name<-}
\alias{name<-}
\title{Set the project name}
\usage{
name(x) <- value
}
\arguments{
\item{x}{A SyncroSim \code{\link{Project}} object.}

\item{value}{The new project name.}
}
\description{
Set the name of a SyncroSim project.
}

