% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{name<-}
\alias{name<-}
\alias{name<-,character-method}
\alias{name<-,SsimLibrary-method}
\alias{name<-,Project-method}
\alias{name<-,Scenario-method}
\title{Set ssimObject name.}
\usage{
name(ssimObject) <- value

\S4method{name}{character}(ssimObject) <- value

\S4method{name}{SsimLibrary}(ssimObject) <- value

\S4method{name}{Project}(ssimObject) <- value

\S4method{name}{Scenario}(ssimObject) <- value
}
\arguments{
\item{ssimObject}{An object of class \code{\link{Session}}, \code{\link{Project}}, 
or \code{\link{SsimLibrary}}.}

\item{value}{A character string, the new name.}
}
\value{
The updated ssim Object.
}
\description{
Set the name of a \code{\link{SsimLibrary}}, \code{\link{Project}}, or
 \code{\link{Scenario}}.
}
\examples{
\donttest{
temp_dir <- tempdir()
mySession <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"), session = mySession)
myProject <- project(myLibrary)
myScenario <- scenario(myProject)

name(myScenario) <- "Scenario Test"
}

}
