% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{name<-}
\alias{name<-}
\alias{name<-,character-method}
\alias{name<-,SsimLibrary-method}
\alias{name<-,Project-method}
\alias{name<-,Scenario-method}
\title{Set ssimObject name.}
\usage{
name(ssimObject) <- value

\S4method{name}{character}(ssimObject) <- value

\S4method{name}{SsimLibrary}(ssimObject) <- value

\S4method{name}{Project}(ssimObject) <- value

\S4method{name}{Scenario}(ssimObject) <- value
}
\arguments{
\item{ssimObject}{Scenario/Project/SsimLibrary}

\item{value}{The new name.}
}
\description{
Set the name of a SyncroSim Project, Scenario or Library
}
