% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{name}
\alias{name}
\title{The name of a SyncroSim project or scenario.}
\usage{
name(ssimObject)
}
\arguments{
\item{ssimObject}{Project/Scenario.}
}
\description{
The name of a SyncroSim Project/Scenario.
}

