% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{name}
\alias{name}
\alias{name,character-method}
\alias{name,SsimLibrary-method}
\alias{name,Scenario-method}
\alias{name,Project-method}
\title{Retrieves the name of a SsimLibrary, Project or Scenario.}
\usage{
name(ssimObject)

\S4method{name}{character}(ssimObject)

\S4method{name}{SsimLibrary}(ssimObject)

\S4method{name}{Scenario}(ssimObject)

\S4method{name}{Project}(ssimObject)
}
\arguments{
\item{ssimObject}{An object of class \code{\link{Scenario}}, \code{\link{Project}}, 
or \code{\link{SsimLibrary}}.}
}
\value{
A character string: the name of the ssimObject.
}
\description{
Retrieves the name of a \code{\link{SsimLibrary}}, \code{\link{Project}} or 
\code{\link{Scenario}}.
}
\examples{
\donttest{
temp_dir <- tempdir()
mySession <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"), session = mySession)
myProject <- project(myLibrary, project = "Definitions")
myScenario <- scenario(myProject, scenario = "My Scenario")

name(myLibrary)
name(myProject)
name(myScenario)
}

}
