% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{name}
\alias{name}
\alias{name,character-method}
\alias{name,SsimLibrary-method}
\alias{name,Scenario-method}
\alias{name,Project-method}
\title{The name of a SsimLibrary, Project or Scenario.}
\usage{
name(ssimObject)

\S4method{name}{character}(ssimObject)

\S4method{name}{SsimLibrary}(ssimObject)

\S4method{name}{Scenario}(ssimObject)

\S4method{name}{Project}(ssimObject)
}
\arguments{
\item{ssimObject}{SsimLibrary, Project, or Scenario.}
}
\value{
Character string: the name of the ssimObject.
}
\description{
Retrieves the name of a \code{\link{SsimLibrary}}, \code{\link{Project}} or \code{\link{Scenario}}.
}
