% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{name}
\alias{name}
\title{The name of a SyncroSim library, project or scenario.}
\usage{
name(ssimObject)
}
\arguments{
\item{ssimObject}{SsimLibrary, Project, or Scenario.}
}
\description{
The name of an SsimLibrary, Project or Scenario.
}

