% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owner.R
\name{owner<-}
\alias{owner<-}
\alias{owner<-,character-method}
\alias{owner<-,SsimObject-method}
\title{Set the owner of a SsimLibrary, Project or Scenario.}
\usage{
owner(ssimObject) <- value

\S4method{owner}{character}(ssimObject) <- value

\S4method{owner}{SsimObject}(ssimObject) <- value
}
\arguments{
\item{ssimObject}{An object of class \code{\link{Session}}, \code{\link{Project}},
or \code{\link{SsimLibrary}}.}

\item{value}{A character string, the new owner.}
}
\value{
The updated ssimObject.
}
\description{
Set the owner of a \code{\link{SsimLibrary}}, \code{\link{Project}} or
\code{\link{Scenario}}.
}
\examples{
\donttest{
temp_dir <- tempdir()
mySession <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"), session = mySession)
myProject <- project(myLibrary, project = "Definitions")
myScenario <- scenario(myProject, scenario = "My Scenario")

owner(myLibrary) <- "Apex RMS"
}

}
