% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{owner<-}
\alias{owner<-}
\title{Set the owner of an SsimLibrary/Project/Scenario.}
\usage{
owner(ssimObject) <- value
}
\arguments{
\item{ssimObject}{Scenario/Project/SsimLibrary}

\item{value}{The new owner.}
}
\description{
Set the owner of an SsimLibrary/Project/Scenario.
}

