% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owner.R
\name{owner<-}
\alias{owner<-}
\alias{owner<-,character-method}
\alias{owner<-,SsimObject-method}
\title{Set the owner of an SsimLibrary/Project/Scenario.}
\usage{
owner(ssimObject) <- value

\S4method{owner}{character}(ssimObject) <- value

\S4method{owner}{SsimObject}(ssimObject) <- value
}
\arguments{
\item{ssimObject}{Scenario/Project/SsimLibrary}

\item{value}{The new owner.}
}
\value{
The updated ssimObject.
}
\description{
Set the owner of an SsimLibrary/Project/Scenario.
}
