% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owner.R
\name{owner}
\alias{owner}
\alias{owner,character-method}
\alias{owner,SsimLibrary-method}
\alias{owner,Project-method}
\alias{owner,Scenario-method}
\title{Retrieves the owner of a SsimLibrary, Project or Scenario.}
\usage{
owner(ssimObject)

\S4method{owner}{character}(ssimObject)

\S4method{owner}{SsimLibrary}(ssimObject)

\S4method{owner}{Project}(ssimObject)

\S4method{owner}{Scenario}(ssimObject)
}
\arguments{
\item{ssimObject}{An object of class \code{\link{Session}}, \code{\link{Project}}, 
or \code{\link{SsimLibrary}}.}
}
\value{
A character string: the owner of the ssimObject.
}
\description{
Retrieves the owner of a \code{\link{SsimLibrary}}, \code{\link{Project}} or 
\code{\link{Scenario}}.
}
\examples{
\donttest{
temp_dir <- tempdir()
mySession <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"), session = mySession)
myProject <- project(myLibrary, project = "Definitions")
myScenario <- scenario(myProject, scenario = "My Scenario")

owner(myLibrary)
owner(myProject)
owner(myScenario)
}

}
