% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owner.R
\docType{methods}
\name{owner}
\alias{owner}
\alias{owner,Project-method}
\alias{owner,Scenario-method}
\alias{owner,SsimLibrary-method}
\title{The owner of a SsimLibrary/Project/Scenario.}
\usage{
owner(ssimObject)

\S4method{owner}{SsimLibrary}(ssimObject)

\S4method{owner}{Project}(ssimObject)

\S4method{owner}{Scenario}(ssimObject)
}
\arguments{
\item{ssimObject}{SsimLibrary/Project/Scenario.}
}
\description{
The owner of an SsimLibrary/ProjectScenario
}
\section{Methods (by class)}{
\itemize{
\item \code{SsimLibrary}: Owner of an SsimLibrary.

\item \code{Project}: Owner of a Project.

\item \code{Scenario}: Owner of a Scenario.
}}

