% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owner.R
\name{owner}
\alias{owner}
\alias{owner,character-method}
\alias{owner,SsimLibrary-method}
\alias{owner,Project-method}
\alias{owner,Scenario-method}
\title{The owner of a SsimLibrary/Project/Scenario.}
\usage{
owner(ssimObject)

\S4method{owner}{character}(ssimObject)

\S4method{owner}{SsimLibrary}(ssimObject)

\S4method{owner}{Project}(ssimObject)

\S4method{owner}{Scenario}(ssimObject)
}
\arguments{
\item{ssimObject}{SsimLibrary/Project/Scenario.}
}
\value{
A character string: the owner of the ssimObject.
}
\description{
Retrieves the owner of a SsimLibrary/ProjectScenario.
}
