% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owner.R
\name{owner}
\alias{owner}
\title{The owner of a SsimLibrary/Project/Scenario.}
\usage{
owner(ssimObject)
}
\arguments{
\item{ssimObject}{SsimLibrary/Project/Scenario.}
}
\description{
The owner of an SsimLibrary/ProjectScenario
}

