% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{package}
\alias{package}
\alias{package,character-method}
\alias{package,missingOrNULL-method}
\alias{package,Session-method}
\alias{package,SsimLibrary-method}
\title{Retrieves the installed or available packages}
\usage{
package(ssimObject = NULL, installed = TRUE)

\S4method{package}{character}(ssimObject = NULL, installed = TRUE)

\S4method{package}{missingOrNULL}(ssimObject = NULL, installed = TRUE)

\S4method{package}{Session}(ssimObject = NULL, installed = TRUE)

\S4method{package}{SsimLibrary}(ssimObject)
}
\arguments{
\item{ssimObject}{An object of class \code{\link{Session}} or 
\code{\link{SsimLibrary}}.}

\item{installed}{Logical or Character. `TRUE` to list installed packages, 
`FALSE` to list available packages, and "BASE" to list installed base 
packages. Default is `TRUE`}
}
\value{
Returns a \code{data.frame} of packages installed.
}
\description{
Retrieves the packages installed or available for this version of SyncroSim.
}
\examples{
\donttest{
temp_dir <- tempdir()
mySession <- session()

package(mySession)
}
\donttest{
temp_dir <- tempdir()
mySession <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"), session = mySession)

package(myLibrary)
}

}
