% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{package}
\alias{package}
\alias{package,missingOrNULL-method}
\alias{package,character-method}
\alias{package,Session-method}
\title{Installed or available packages}
\usage{
package(session, installed = TRUE)

\S4method{package}{missingOrNULL}(session, installed = TRUE)

\S4method{package}{character}(session, installed = TRUE)

\S4method{package}{Session}(session, installed = TRUE)
}
\arguments{
\item{session}{Session.}

\item{installed}{Logical. `TRUE` to list installed packages and `FALSE` to list 
available packages}
}
\value{
A \code{data.frame} of packages installed.
}
\description{
Packages or installed or available for this version of SyncroSim.
}
