% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{package}
\alias{package}
\alias{package,missingOrNULL-method}
\alias{package,character-method}
\alias{package,Session-method}
\title{Installed or available packages}
\usage{
package(session, installed = T)

\S4method{package}{missingOrNULL}(session, installed = T)

\S4method{package}{character}(session, installed = T)

\S4method{package}{Session}(session, installed = T)
}
\arguments{
\item{session}{Session.}

\item{installed}{Logical.  True to list installed packages and False to list available pacakges.}
}
\value{
A dataframe of packages
}
\description{
Packages or installed or available for this version of SyncroSim.
}
