% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{parentId}
\alias{parentId}
\title{The parent scenario id of a SyncroSim Scenario.}
\usage{
parentId(scenario)
}
\arguments{
\item{scenario}{A Scenario object.}
}
\description{
The id of the parent of a SyncroSim results scenario.
NA if scenario is not a results scenario.
}

