% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentId.R
\name{parentId}
\alias{parentId}
\alias{parentId,character-method}
\alias{parentId,Scenario-method}
\title{The parent scenario id of a SyncroSim Scenario.}
\usage{
parentId(scenario)

\S4method{parentId}{character}(scenario)

\S4method{parentId}{Scenario}(scenario)
}
\arguments{
\item{scenario}{A Scenario object.}
}
\value{
An integer id of the parent scenario. If the input scenario does not have a
parent, the function returns `NA`.
}
\description{
Retrives the id of the parent of a SyncroSim results scenario.
}
