% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentId.R
\name{parentId}
\alias{parentId}
\alias{parentId,character-method}
\alias{parentId,Scenario-method}
\title{The parent scenario id of a SyncroSim Scenario.}
\usage{
parentId(scenario)

\S4method{parentId}{character}(scenario)

\S4method{parentId}{Scenario}(scenario)
}
\arguments{
\item{scenario}{A Scenario object.}
}
\value{
An integer id of the parent scenario.
}
\description{
The id of the parent of a SyncroSim results scenario.
NA if scenario is not a results scenario.
}
