% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{printCmd}
\alias{printCmd}
\title{Get printCmd a \code{\link{Session}}.}
\usage{
printCmd(session = NULL)
}
\arguments{
\item{session}{Session or character. A SyncroSim \code{\link{Session}} object or path to a session. If NULL, the default session will be used.}
}
\description{
Get printCmd a \code{\link{Session}}.
}

