% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printCmd.R
\name{printCmd}
\alias{printCmd}
\alias{printCmd,Session-method}
\alias{printCmd,missingOrNULLOrChar-method}
\title{Retrieves a printCmd setting of a Session.}
\usage{
printCmd(session = NULL)

\S4method{printCmd}{Session}(session = NULL)

\S4method{printCmd}{missingOrNULLOrChar}(session = NULL)
}
\arguments{
\item{session}{Session or character. An object of class \code{\link{Session}}
or path to a session. If NULL, the default session will be used.}
}
\value{
Returns a logical value: \code{TRUE} if the session is configured to print commands and
\code{FALSE} if it is not.
}
\description{
Retrieves a printCmd setting of a \code{\link{Session}} object. The printCmd
setting configures a session for printing commands sent to the console.
}
\examples{
\donttest{
temp_dir <- tempdir()
mySession <- session()

printCmd(mySession)
}

}
