% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printCmd.R
\name{printCmd}
\alias{printCmd}
\title{Get printCmd of a Session.}
\usage{
printCmd(session = NULL)
}
\arguments{
\item{session}{Session or character. A Session object or path to a session. If NULL, the default session will be used.}
}
\value{
logical.
}
\description{
Get printCmd setting of a Session object.
}

