% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printCmd.R
\name{printCmd}
\alias{printCmd}
\alias{printCmd,Session-method}
\alias{printCmd,missingOrNULLOrChar-method}
\title{Get printCmd of a Session.}
\usage{
printCmd(session = NULL)

\S4method{printCmd}{Session}(session = NULL)

\S4method{printCmd}{missingOrNULLOrChar}(session = NULL)
}
\arguments{
\item{session}{Session or character. A Session object or path to a session. 
If NULL, the default session will be used.}
}
\value{
Returns a logical value: `TRUE` is the session is configured to print commands and 
`FALSE` if it is not.
}
\description{
Retrives a printCmd setting of a Session object.
}
