% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAClassDefinitions.R
\docType{class}
\name{Project-class}
\alias{Project-class}
\alias{Project}
\title{SyncroSim Project class}
\description{
\code{Project} object representing a SyncroSim Project.
}
\section{Slots}{

\describe{
\item{\code{session}}{The Session associated with the Project's Library.}

\item{\code{filepath}}{The path to the Project's Library on disk.}

\item{\code{datasheetNames}}{Names and scopes of datasheets in the Project's Library.}

\item{\code{projectId}}{The Project id.}
}}

\seealso{
See \code{\link{project}} for options when creating or loading a SyncroSim Project.
}
