% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAClassDefinitions.R
\docType{class}
\name{Project-class}
\alias{Project}
\alias{Project-class}
\title{SyncroSim Project class}
\description{
\code{Project} object representing a SyncroSim Project.
}
\section{Slots}{

\describe{
\item{\code{session}}{The session associated with the library.}

\item{\code{filepath}}{The path to the library on disk.}

\item{\code{datasheetNames}}{Names and scopes of datasheets in the library.}

\item{\code{projectId}}{The project id}
}}
\seealso{
See \code{\link{project}} for options when creating or loading an SyncroSim Project.
}

