% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Create or open a project.}
\usage{
project(ssimLibrary, name = NULL, id = NULL)
}
\arguments{
\item{ssimLibrary}{An SSimLibrary object, representing the library that contains the project.}

\item{name}{The project name. At least one of name and id must be provided to identify the project.}

\item{id}{The project id. At least one of name and id must be provided to identify the project.}
}
\value{
A \code{Project} object representing a SyncroSim project.
}
\description{
Creates or opens an \code{\link{Project}} object representing a SyncroSim project.
}
\examples{
# Create a project in the default library.
myLib = ssimLibrary(model="stsim",name="stsim")
myProject = project(myLib,name="firstProject")
}

