% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{projectId}
\alias{projectId}
\title{The projectId of a SyncroSim project or scenario.}
\usage{
projectId(ssimObject)
}
\arguments{
\item{ssimObject}{Project/Scenario.}
}
\description{
The projectId of a SyncroSim Project or Scenario.
}

