% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectId.R
\name{projectId}
\alias{projectId}
\alias{projectId,character-method}
\alias{projectId,Project-method}
\alias{projectId,Scenario-method}
\title{The projectId of a SyncroSim Project or Scenario.}
\usage{
projectId(ssimObject)

\S4method{projectId}{character}(ssimObject)

\S4method{projectId}{Project}(ssimObject)

\S4method{projectId}{Scenario}(ssimObject)
}
\arguments{
\item{ssimObject}{Project or Scenario.}
}
\value{
Returns an integer project id.
}
\description{
Retrieves the projectId of a SyncroSim \code{\link{Project}} or \code{\link{Scenario}}.
}
