% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectId.R
\docType{methods}
\name{projectId}
\alias{projectId}
\alias{projectId,Project-method}
\alias{projectId,Scenario-method}
\title{The projectId of a SyncroSim project or scenario.}
\usage{
projectId(ssimObject)

\S4method{projectId}{Project}(ssimObject)

\S4method{projectId}{Scenario}(ssimObject)
}
\arguments{
\item{ssimObject}{Project/Scenario.}
}
\value{
An integer project id.
}
\description{
The projectId of a SyncroSim Project or Scenario.
}
\section{Methods (by class)}{
\itemize{
\item \code{Project}: The id of a Project.

\item \code{Scenario}: The projectId of a Scenario.
}}

