% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectId.R
\name{projectId}
\alias{projectId}
\alias{projectId,character-method}
\alias{projectId,Project-method}
\alias{projectId,Scenario-method}
\title{The projectId of a SyncroSim project or scenario.}
\usage{
projectId(ssimObject)

\S4method{projectId}{character}(ssimObject)

\S4method{projectId}{Project}(ssimObject)

\S4method{projectId}{Scenario}(ssimObject)
}
\arguments{
\item{ssimObject}{Project/Scenario.}
}
\value{
An integer project id.
}
\description{
The projectId of a SyncroSim Project or Scenario.
}
