% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimLibrary.R
\name{projects}
\alias{projects}
\title{The projects in a SyncroSim library.}
\usage{
projects(x, names = F)
}
\arguments{
\item{x}{An SSimLibrary object, or a Project or Scenario associated with a Library}

\item{names}{If FALSE, a list of \code{\link{Project}} objects is returned. If TRUE returns a dataframe containing the name and id of each project.}
}
\value{
By default returns a list of projects identified by the project id. Each element of the list contains a SyncroSim Project object. If names=T, returns a dataframe containing the name and id of each project.
}
\description{
Get a list of projects in a SyncroSim library.
}
\examples{
myProjects = projects(ssimLibrary(model="stsim",name="stsim"))
}

