% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{readOnly<-}
\alias{readOnly<-}
\title{Set the read/write status of an SsimLibrary/Project/Scenario.}
\usage{
readOnly(ssimObject) <- value
}
\arguments{
\item{ssimObject}{Scenario/Project/SsimLibrary}

\item{value}{Logical. If T the ssimObject will be read-only.}
}
\description{
Set the read-only status of an SsimLibrary/Project/Scenario. 
Applies to child objects if ssimObject is an SsimLibrary or Project.
}

