% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOnly.R
\name{readOnly<-}
\alias{readOnly<-}
\alias{readOnly<-,character-method}
\alias{readOnly<-,SsimObject-method}
\title{Set the read/write status of a SsimLibrary, Project or Scenario.}
\usage{
readOnly(ssimObject) <- value

\S4method{readOnly}{character}(ssimObject) <- value

\S4method{readOnly}{SsimObject}(ssimObject) <- value
}
\arguments{
\item{ssimObject}{SsimLibrary, Project, or Scenario.}

\item{value}{Logical. If T the ssimObject will be read-only.}
}
\value{
The updated ssimObject.
}
\description{
Set the read-only status of a \code{\link{SsimLibrary}}, \code{\link{Project}} or \code{\link{Scenario}}.
Applies to child objects if ssimObject is a SsimLibrary or Project.
}
