% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOnly.R
\name{readOnly<-}
\alias{readOnly<-}
\alias{readOnly<-,character-method}
\alias{readOnly<-,SsimObject-method}
\title{Set the read/write status of an SsimLibrary/Project/Scenario.}
\usage{
readOnly(ssimObject) <- value

\S4method{readOnly}{character}(ssimObject) <- value

\S4method{readOnly}{SsimObject}(ssimObject) <- value
}
\arguments{
\item{ssimObject}{Scenario/Project/SsimLibrary}

\item{value}{Logical. If T the ssimObject will be read-only.}
}
\value{
The updated ssimObject.
}
\description{
Set the read-only status of an SsimLibrary/Project/Scenario.
Applies to child objects if ssimObject is an SsimLibrary or Project.
}
