% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{readOnly}
\alias{readOnly}
\title{The write status of a Scenario}
\usage{
readOnly(x)
}
\arguments{
\item{x}{An Scenario object.}
}
\value{
TRUE or FALSE
}
\description{
Whether or not the scenario is readOnly
}

