% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{readOnly}
\alias{readOnly}
\title{Read-only status of an SsimLibrary/Project/Scenario.}
\usage{
readOnly(ssimObject)
}
\arguments{
\item{ssimObject}{SsimLibrary/Project/Scenario.}
}
\description{
Whether or not an SsimLibrary/ProjectScenario is read-only.
}

