% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOnly.R
\name{readOnly}
\alias{readOnly}
\alias{readOnly,character-method}
\alias{readOnly,SsimLibrary-method}
\alias{readOnly,Project-method}
\alias{readOnly,Scenario-method}
\title{Retrieves the Read-only status of a SsimLibrary, Project or Scenario.}
\usage{
readOnly(ssimObject)

\S4method{readOnly}{character}(ssimObject)

\S4method{readOnly}{SsimLibrary}(ssimObject)

\S4method{readOnly}{Project}(ssimObject)

\S4method{readOnly}{Scenario}(ssimObject)
}
\arguments{
\item{ssimObject}{An object of class \code{\link{Session}}, \code{\link{Project}}, 
or \code{\link{SsimLibrary}}.}
}
\value{
Returns a logical value: `TRUE` if the ssimObject is read only and `FALSE`
otherwise.
}
\description{
Whether or not a \code{\link{SsimLibrary}}, \code{\link{Project}} or 
\code{\link{Scenario}} is read-only is read-only.
}
\examples{
\donttest{
temp_dir <- tempdir()
mySession <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"), session = mySession)
myProject <- project(myLibrary, project = "Definitions")
myScenario <- scenario(myProject, scenario = "My Scenario")

readOnly(myLibrary)
readOnly(myProject)
readOnly(myScenario)
}

}
