% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOnly.R
\docType{methods}
\name{readOnly}
\alias{readOnly}
\alias{readOnly,Project-method}
\alias{readOnly,Scenario-method}
\alias{readOnly,SsimLibrary-method}
\alias{readOnly,character-method}
\title{Read-only status of an SsimLibrary/Project/Scenario.}
\usage{
readOnly(ssimObject)

\S4method{readOnly}{character}(ssimObject)

\S4method{readOnly}{SsimLibrary}(ssimObject)

\S4method{readOnly}{Project}(ssimObject)

\S4method{readOnly}{Scenario}(ssimObject)
}
\arguments{
\item{ssimObject}{SsimLibrary/Project/Scenario.}
}
\value{
logical.
}
\description{
Whether or not an SsimLibrary/ProjectScenario is read-only.
}

