% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/scenario.R
\name{readOnly}
\alias{readOnly}
\title{Read-only status of an SsimLibrary/Project/Scenario.}
\usage{
readOnly(x)

readOnly(x)
}
\arguments{
\item{x}{An Scenario object.}

\item{ssimObject}{SsimLibrary/Project/Scenario.}
}
\value{
TRUE or FALSE
}
\description{
Whether or not an SsimLibrary/ProjectScenario is read-only.

Whether or not the scenario is readOnly
}

