% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{removeModules<-}
\alias{removeModules<-}
\title{Remove modules}
\usage{
removeModules(x) <- value
}
\arguments{
\item{x}{A SyncroSim \code{\link{Session}} object.}

\item{value}{A module or vector of modules to remove. \code{\link{modules()}} for options.}
}
\description{
Remove module or modules to this version of SyncroSim.
TO DO: Doesn't currently do anything. Need console commands for adding and removing modules.
}

