% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removePackage.R
\name{removePackage}
\alias{removePackage}
\alias{removePackage,ANY,character-method}
\alias{removePackage,ANY,missingOrNULL-method}
\alias{removePackage,ANY,Session-method}
\title{Removes a package from your SyncroSim installation.}
\usage{
removePackage(name, session = NULL, force = FALSE)

\S4method{removePackage}{ANY,character}(name, session = NULL, force = FALSE)

\S4method{removePackage}{ANY,missingOrNULL}(name, session = NULL, force = FALSE)

\S4method{removePackage}{ANY,Session}(name, session = NULL, force = FALSE)
}
\arguments{
\item{name}{Character. The name of the package to remove}

\item{session}{An object of class \code{\link{Session}}.}

\item{force}{Logical. If TRUE, remove without requiring confirmation from
the user. Default is FALSE.}
}
\value{
This function invisibly returns \code{TRUE} upon success (i.e.successful
removal) and \code{FALSE} upon failure.
}
\description{
Removes a package from your SyncroSim installation.
}
\examples{
\donttest{
temp_dir <- tempdir()
mySession <- session()

removePackage("stsim", mySession, force = FALSE)
}

}
