% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removePackage.R
\name{removePackage}
\alias{removePackage}
\alias{removePackage,character-method}
\alias{removePackage,SsimLibrary-method}
\title{Removes SyncroSim package(s)}
\usage{
removePackage(ssimLibrary, name)

\S4method{removePackage}{character}(ssimLibrary, name)

\S4method{removePackage}{SsimLibrary}(ssimLibrary, name)
}
\arguments{
\item{ssimLibrary}{\code{\link{SsimLibrary}} object}

\item{name}{character string or vector of package name(s)}
}
\value{
This function invisibly returns \code{TRUE} upon success (i.e.successful
removal of the package) or \code{FALSE} upon failure.
}
\description{
Removes package(s) from a \code{\link{SsimLibrary}}.
}
\examples{
\donttest{
# Install "stsim" SyncroSim package
installPackage("stsim")

# Specify file path and name of new SsimLibrary
myLibraryName <- file.path(tempdir(), "testlib")

# Set up a SyncroSim Session, SsimLibrary, and Project
mySession <- session()
myLibrary <- ssimLibrary(name = myLibraryName, session = mySession, 
                         package = "stsim")

# Add package
addPackage(myLibrary, c("stsimsf"))
packages(myLibrary)

# Remove package
removePackage(myLibrary, c("stsimsf"))
packages(myLibrary)
}

}
\seealso{
\code{\link{packages}}
}
