% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.R
\name{restore}
\alias{restore}
\alias{restore,character-method}
\title{Restore Library}
\usage{
restore(ssimLibraryBackup, folder = NULL, session = NULL)

\S4method{restore}{character}(ssimLibraryBackup, folder = NULL, session = NULL)
}
\arguments{
\item{ssimLibraryBackup}{SsimLibrary backup file or path to a library backup file}

\item{folder}{Optional folder to restore the library into. If NULL, restores to default location.}

\item{session}{SyncroSim session}
}
\value{
"Library restored" or "Library restoration failed".
}
\description{
Restores a SyncroSim library from a backup file.
}
\examples{
\dontrun{
# Specify file path and name of SsimLibrary backup file
myLibraryBackupName <- file.path(tempdir(), "testlib.ssimbak")

# Set up a SyncroSim Session and create SsimLibrary from backup file
mySession <- session()
myLibrary <- restore(ssimLibraryBackup = myLibraryBackupName, session = mySession)

}

}
