% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsyncrosim-package.R
\docType{package}
\name{rsyncrosim-package}
\alias{rsyncrosim-package}
\title{rsyncrosim: The R Interface to 'SyncroSim'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

'SyncroSim' is a generalized framework for managing scenario-based datasets (\url{https://syncrosim.com/}). 'rsyncrosim' provides an interface to 'SyncroSim'. Simulation models can be added to 'SyncroSim' in order to transform these datasets, taking advantage of general features such as defining scenarios of model inputs, running Monte Carlo simulations, and summarizing model outputs. 'rsyncrosim' requires 'SyncroSim' 2.3.5 or higher (API documentation: \url{https://docs.syncrosim.com/}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://syncrosim.github.io/rsyncrosim/}
  \item Report bugs at \url{https://github.com/syncrosim/rsyncrosim/issues/}
}

}
\author{
\strong{Maintainer}: Katie Birchard \email{katie.birchard@apexrms.com}

Authors:
\itemize{
  \item Colin Daniel \email{colin.daniel@apexrms.com}
  \item Josie Hughes \email{josie.hughes@canada.ca}
  \item Valentin Lucet
  \item Alex Embrey
  \item Leonardo Frid
  \item Tabitha Kennedy
  \item Shreeram Senthivasan
}

Other contributors:
\itemize{
  \item ApexRMS [copyright holder]
}

}
\keyword{internal}
