% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{run}
\alias{run}
\title{Run scenarios}
\usage{
run(x, scenario = NULL, onlyIds = F)
}
\arguments{
\item{x}{One or more SSimLibrary, Projects or Scenario objects. Or the path to a library on disk.}

\item{scenario}{One or more scenario objects, names or ids.}

\item{onlyIds}{If FALSE (default) result Scenario objects are returned. If TRUE (faster) result scenario ids are returned.}
}
\value{
A named list of result Scenario objects or ids. The name is the parent scenario for each result.
}
\description{
Run one or more SyncroSim scenarios
}
\examples{

}

