% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLog.R
\docType{methods}
\name{runLog}
\alias{runLog}
\alias{runLog,Scenario-method}
\alias{runLog,character-method}
\title{The runLog of a result Scenario.}
\usage{
runLog(scenario)

\S4method{runLog}{character}(scenario)

\S4method{runLog}{Scenario}(scenario)
}
\arguments{
\item{scenario}{A Scenario object.}
}
\value{
Character string of the run log.
}
\description{
The runLog of a result Scenario
}

