% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLog.R
\name{runLog}
\alias{runLog}
\alias{runLog,character-method}
\alias{runLog,Scenario-method}
\title{The runLog of a result Scenario}
\usage{
runLog(scenario)

\S4method{runLog}{character}(scenario)

\S4method{runLog}{Scenario}(scenario)
}
\arguments{
\item{scenario}{A Scenario object.}
}
\value{
Character string of the run log.
}
\description{
The runLog of a result Scenario.
}
