% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLog.R
\name{runLog}
\alias{runLog}
\title{The runLog of a result Scenario.}
\usage{
runLog(scenario)
}
\arguments{
\item{scenario}{A Scenario object.}
}
\value{
Character string of the run log.
}
\description{
The runLog of a result Scenario
}

