% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLog.R
\name{runLog}
\alias{runLog}
\alias{runLog,character-method}
\alias{runLog,Scenario-method}
\title{Retrieves the run log of a result Scenario}
\usage{
runLog(scenario)

\S4method{runLog}{character}(scenario)

\S4method{runLog}{Scenario}(scenario)
}
\arguments{
\item{scenario}{A \code{\link{Scenario}} object.}
}
\value{
Returns a character string: the run log for a result scenario.
}
\description{
Retrieves the run log of a result Scenario.
}
\examples{
\donttest{
temp_dir <- tempdir()
mySession <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"), session = mySession)
myProject <- project(myLibrary)
myScenario <- scenario(myProject)

# Note: the scenario must be a result scenario. This example only shows the 
# how to use the function with a Scenario object.
runLog(myScenario)
}

}
