% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLog.R
\name{runLog}
\alias{runLog}
\alias{runLog,character-method}
\alias{runLog,Scenario-method}
\title{Retrieves the run log of a result Scenario}
\usage{
runLog(scenario)

\S4method{runLog}{character}(scenario)

\S4method{runLog}{Scenario}(scenario)
}
\arguments{
\item{scenario}{A \code{\link{Scenario}} object.}
}
\value{
Returns a character string: the run log for a result scenario.
}
\description{
Retrieves the run log of a result Scenario.
}
\examples{
\dontrun{
addPackage("helloworldEnhanced")
temp_dir <- tempdir()
mySession <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"),
                         session = mySession, package = "helloworldEnhanced",
                         template = "example-library")
myProject <- project(myLibrary, project = "Definitions")
myScenario <- scenario(myProject, scenario = "My Scenario")

resultScenario <- run(myScenario)

runLog(resultScenario)
}

}
