% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveDatasheet.R
\docType{methods}
\name{saveDatasheet}
\alias{saveDatasheet}
\alias{saveDatasheet,character-method}
\alias{saveDatasheet,SsimObject-method}
\title{Save datasheet(s)}
\usage{
saveDatasheet(ssimObject, data, name = NULL, append = NULL,
  forceElements = F, force = F, breakpoint = F, import = T,
  path = NULL)

\S4method{saveDatasheet}{character}(ssimObject, data, name = NULL,
  append = NULL, forceElements = F, force = F, breakpoint = F,
  import = T, path = NULL)

\S4method{saveDatasheet}{SsimObject}(ssimObject, data, name = NULL,
  append = NULL, forceElements = F, force = F, breakpoint = F,
  import = T, path = NULL)
}
\arguments{
\item{ssimObject}{SsimLibrary/Project/Scenario.}

\item{data}{A dataframe, named vector, or list of these. One or more datasheets to load.}

\item{name}{character or vector of these. The name(s) of the datasheet(s) to be saved. If a vector of names is provided, then a list must be provided for the data argument. Names provided here will override those provided with data argument's list.}

\item{append}{logical. If TRUE, data will be appended to the datasheet if possible, otherwise current values will be overwritten by data. See details for behaviour when append=T. Default TRUE for project/library-scope datasheets, and FALSE for scenario-scope datasheets.}

\item{forceElements}{logical. If FALSE (default) a single return message will be returns as a character string. Otherwise it will be returned in a list.}

\item{force}{logical. If datasheet scope is project/library, and append=F, datasheet will be deleted before loading the new data. This can also delete other definitions and results, so user will be prompted for approval unless force=T.}

\item{breakpoint}{Set to TRUE when modifying datasheets in a breakpoint function.}

\item{import}{logical. Set to TRUE to import the data after saving.}

\item{path}{character.  An optional output path.}
}
\value{
A success or failure message, or a list of these.
}
\description{
Saves datasheets to a SsimLibrary/Project/Scenario.
}
\details{
Cautionary note re append=F: Deleting project and library level datasheets that contain lookups will also delete other definitions and results that rely on these lookups. 

ssimObject/project/scenario should identify a single ssimObject.


There are 2 circumstances in which data will not be appended even if append=T:
\itemize{
  \item New data will not be appended if it is redundant with existing data, and the table does not allow redundancy.
  \item Old data will be replaced by new data if the datasheet allows only a single row.
}
}
