% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAClassDefinitions.R
\docType{class}
\name{Scenario-class}
\alias{Scenario}
\alias{Scenario-class}
\title{SyncroSim Scenario class}
\description{
\code{Scenario} object representing a SyncroSim Project.
}
\section{Slots}{

\describe{
\item{\code{session}}{The session associated with the library.}

\item{\code{filepath}}{The path to the library on disk.}

\item{\code{datasheetNames}}{Names and scope of all datasheets in library.}

\item{\code{pid}}{The project id.}

\item{\code{name}}{The scenario name.}

\item{\code{id}}{The scenario id.}

\item{\code{parentId}}{For a result scenario, this is the id of the parent scenario. 0 indicates this is not a result scenario.}

\item{\code{breakpoints}}{An (optional) list of Breakpoint objects. See ?breakpoints for details.}
}}
\seealso{
See \code{\link{scenario}} for options when creating or loading an SyncroSim Scenario.
}

