% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAClassDefinitions.R
\docType{class}
\name{Scenario-class}
\alias{Scenario-class}
\alias{Scenario}
\title{SyncroSim Scenario class}
\description{
\code{Scenario} object representing a SyncroSim Scenario. A Scenario is the 
lowest level of organization in the SyncroSim workflow, and is often used 
to isolate information on a single Datasheet.
}
\section{Slots}{

\describe{
\item{\code{session}}{\code{\link{Session}} object. The Session associated with the 
Scenario}

\item{\code{filepath}}{character string. The path to the Scenario's SsimLibrary on disk}

\item{\code{datasheetNames}}{character string. Names and scope of all Datasheets in
Scenario's SsimLibrary}

\item{\code{projectId}}{integer. The Project id}

\item{\code{scenarioId}}{integer. The Scenario id}

\item{\code{parentId}}{integer. For a result Scenario, this is the id of the parent 
Scenario. 0 indicates this is not a result Scenario}

\item{\code{breakpoints}}{list of Breakpoint objects (optional)}
}}

\seealso{
See \code{\link{scenario}} for options when creating or loading a 
    SyncroSim Scenario.
}
