% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{scenario}
\alias{scenario}
\title{Create or open a scenario}
\usage{
scenario(x, name = NULL, id = NULL, pid = NULL)
}
\arguments{
\item{x}{An SSimLibrary or Project object.}

\item{name}{The scenario name.}

\item{id}{The scenario id.}

\item{pid}{The project id.}
}
\value{
A \code{Scenario} object representing a SyncroSim scenario.
}
\description{
Creates or opens an \code{\link{Scenario}} object representing a SyncroSim scenario.
}
\details{
If x is a Project, the pid argument is ignored, and pid is set to id(x).
\itemize{
  \item {If name/id/pid uniquely identifies an existing scenario: }{Returns the existing Scenario}
  \item {If name/id/pid uniquely identifies more than one existing scenario: }{Error}
  \item {If pid is NULL, and name/id do not uniquely idenfity an existing scenario: }{Error}
  \item {If pid is not NULL, name is NULL, and id/pid do not idenfity an existing scenario: }{Creates a new Scenario called "Scenario". The id argument is ignored, as SyncroSim automatically assigns an id.}
  \item {If pid is not NULL, name is not NULL, and name/id/pid do not idenfity an existing scenario: }{Creates a new Scenario called <name>. The id argument is ignored, as SyncroSim automatically assigns an id.}
}
}
\examples{
# Create a new default scenario
myLibrary = ssimLibrary(model="stsim",name="stsim")
myProject = project(myLibrary) #If no name is given, creates a project named "Project<ID>".
myScenario = scenario(myProject)

}

