% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{scenario}
\alias{scenario}
\title{Create or open a scenario}
\usage{
scenario(ssimLibrary = NULL, project = NULL, name = NULL, id = NULL,
  create = T, scenarios = NULL)
}
\arguments{
\item{ssimLibrary}{An SSimLibrary object or name, or an object that contains an SSimLibrary. If a name is given, the library will be opened using the default session.}

\item{project}{A Project object, project name, or project id.}

\item{name}{The scenario name.}

\item{id}{The scenario id.}

\item{create}{If TRUE, create scenario if one does not exist. If FALSE, only return an existing scenario}

\item{scenarios}{A dataframe of existing scenarios produced by scenarios(). Use to speed processing.}
}
\value{
A \code{Scenario} object representing a SyncroSim scenario.
}
\description{
Creates or opens an \code{\link{Scenario}} object representing a SyncroSim scenario.
}
\details{
\itemize{
  \item {If name/id/project uniquely identifies an existing scenario: }{Returns the existing Scenario}
  \item {If name/id/project uniquely identifies more than one existing scenario: }{Error}
  \item {If project is NULL, and name/id do not uniquely idenfity an existing scenario: }{Error}
  \item {If project is not NULL, name is NULL, and id/project do not idenfity an existing scenario: }{Creates a new Scenario called "Scenario". The id argument is ignored, as SyncroSim automatically assigns an id.}
  \item {If project is not NULL, name is not NULL, and name/id/project do not idenfity an existing scenario: }{Creates a new Scenario called <name>. The id argument is ignored, as SyncroSim automatically assigns an id.}
}
}
\examples{
# Create a new default scenario
myLibrary = ssimLibrary(model="stsim",name="stsim")
myProject = project(myLibrary) #If no name is given, creates a project named "Project".
myScenario = scenario(myProject)

}

