% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarioId.R
\name{scenarioId}
\alias{scenarioId}
\title{The scenarioId of a scenario.}
\usage{
scenarioId(scenario)
}
\arguments{
\item{scenario}{Scenario.}
}
\value{
integer id.
}
\description{
The scenarioId of a Scenario.
}

