% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarioId.R
\name{scenarioId}
\alias{scenarioId}
\alias{scenarioId,character-method}
\alias{scenarioId,Scenario-method}
\title{The scenarioId of a Scenario.}
\usage{
scenarioId(scenario)

\S4method{scenarioId}{character}(scenario)

\S4method{scenarioId}{Scenario}(scenario)
}
\arguments{
\item{scenario}{Scenario.}
}
\value{
Integer id of the input scenario.
}
\description{
Retrieves the scenarioId of a \code{\link{Scenario}}.
}
