% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarioId.R
\name{scenarioId}
\alias{scenarioId}
\alias{scenarioId,character-method}
\alias{scenarioId,Scenario-method}
\title{The scenarioId of a scenario.}
\usage{
scenarioId(scenario)

\S4method{scenarioId}{character}(scenario)

\S4method{scenarioId}{Scenario}(scenario)
}
\arguments{
\item{scenario}{Scenario.}
}
\value{
integer id.
}
\description{
The scenarioId of a Scenario.
}
