% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarioId.R
\name{scenarioId}
\alias{scenarioId}
\alias{scenarioId,character-method}
\alias{scenarioId,Scenario-method}
\title{Retrieves the scenarioId of a Scenario.}
\usage{
scenarioId(scenario)

\S4method{scenarioId}{character}(scenario)

\S4method{scenarioId}{Scenario}(scenario)
}
\arguments{
\item{scenario}{An object of class \code{\link{Scenario}}.}
}
\value{
Integer id of the input scenario.
}
\description{
Retrieves the scenarioId of a \code{\link{Scenario}}.
}
\examples{
\donttest{
temp_dir <- tempdir()
mySession <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"), session = mySession)
myProject <- project(myLibrary, project = "Definitions")
myScenario <- scenario(myProject, scenario = "My Scenario")

scenarioId(myScenario)
}

}
