% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{scenarios}
\alias{scenarios}
\title{The scenarios in a SyncroSim library or project.}
\usage{
scenarios(x, ...)
}
\arguments{
\item{x}{An SSimLibrary or Project object, or an SSimLibrary name.}

\item{project}{A project name, id, or object.}

\item{names}{If FALSE, a list of \code{\link{Scenario}} objects is returned. If TRUE returns a dataframe containing the name,id and project id of each scenario.}
}
\value{
By default returns a list of scenarios identified by id. Each element of the list contains a SyncroSim Scenario object. If names=T, returns a dataframe containing the name, id, and project id of each scenario.
}
\description{
Get a list of scenarios in a SSimLibrary or Project.
}
\examples{
myScenarios = scenarios(ssimLibrary(model="stsim",name="stsim"))
}

