% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{scenarios}
\alias{scenarios}
\title{The scenarios in a SyncroSim library or project.}
\usage{
scenarios(x, project = NULL, names = F, results = NULL, select = NULL)
}
\arguments{
\item{x}{An SsimLibrary or Project object, or an SsimLibrary name.}

\item{project}{An optional project name, id, or object.}

\item{names}{If FALSE, a list of \code{\link{Scenario}} objects is returned. If TRUE returns a dataframe containing the name,id and project id of each scenario.}

\item{results}{If TRUE only return result scenarios.}

\item{select}{An (optional) vector of scenario ids or names to include}
}
\description{
Get a list of scenarios in a SsimLibrary or Project.
}
\examples{
myScenarios = scenarios(ssimLibrary(model="stsim",name="stsim"))
}

