% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAClassDefinitions.R
\docType{class}
\name{Session-class}
\alias{Session}
\alias{Session-class}
\title{SyncroSim Session class}
\description{
A SyncroSim Session object contains a link to SyncroSim.
\code{SsimLibrary}, \code{Project} and \code{Scenario} objects contain a \code{Session} used to query and modify the object.
}
\section{Slots}{

\describe{
\item{\code{filepath}}{The path to SyncroSim}

\item{\code{silent}}{If TRUE, warnings from the console are ignored. Otherwise they are printed. Default is FALSE.}

\item{\code{printCmd}}{If TRUE, arguments passed to the SyncroSim console are also printed. Helpful for debugging. FALSE by default.}

\item{\code{defaultModel}}{The name of a SyncroSim model type. "stsim" by default.}
}}
\examples{
#TODO - update examples
# Create or load a library using a non-default Session
mySession = session("C:/Program Files/SyncroSim/1/SyncroSim.Console.exe")
myLib = ssimLibrary(name="stsim",session=mySession)
session(myLib)

showMethods(class="Session",where=loadNamespace("rsyncrosim")) #Methods for the Session
filepath(mySession)   # Lists the folder location of syncrosim session
version(mySession)   # Lists the version of syncrosim session
modules(mySession)   # Dataframe of the modules installed with this version of syncrosim.
models(mySsim) # Dataframe of the models installed with this version of syncrosim.

# Add and remove modules
removeModules(mySsim) = "stsim-stock-flow"
is.element("stsim-stock-flow",modules(mySsim)$shortName)
addModules(mySsim) = "C:/Program Files/SyncroSim/1/CorePackages/stockflow.ssimpkg"
addModules(mySsim) = c("C:/Program Files/SyncroSim/1/CorePackages/stockflow.ssimpkg","C:/Program Files/SyncroSim/1/CorePackages/dynmult.ssimpkg")
is.element("stsim-stock-flow",modules(mySsim)$shortName)

# Create or load a library using a default Session
myLib = ssimLibrary(name="stsim")
session(myLib)
}

