% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAClassDefinitions.R
\docType{class}
\name{Session-class}
\alias{Session-class}
\alias{Session}
\title{SyncroSim Session class}
\description{
A SyncroSim Session object contains a link to a SyncroSim installation. 
\code{SsimLibrary}, \code{Project} and \code{Scenario} objects contain a 
\code{Session} used to query and modify the object.
}
\section{Slots}{

\describe{
\item{\code{filepath}}{The path to the SyncroSim installation}

\item{\code{silent}}{If \code{FALSE}, all SyncroSim output with non-zero exit status is 
printed. Helpful for debugging. Default is \code{TRUE}}

\item{\code{printCmd}}{If \code{TRUE}, arguments passed to the SyncroSim console are also 
printed. Helpful for debugging. Default is \code{FALSE}}

\item{\code{condaFilepath}}{The path to the Conda installation. Default is \code{NULL}}
}}

\seealso{
See \code{\link{session}} for options when creating a Session.
}
