% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\docType{class}
\name{Session-class}
\alias{Session}
\alias{Session-class}
\title{SyncroSim Session class}
\description{
A SyncroSim Session object contains a link to SyncroSim.
\code{SSimLibrary}, \code{Project} and \code{Scenario} objects contain a \code{Session} used to query and modify the object.
}
\section{Slots}{

\describe{
\item{\code{filepath}}{The path to SyncroSim.Console.exe.}

\item{\code{silent}}{If TRUE, warnings from the console are ignored. Otherwise they are printed.}
}}
\examples{
#create or load a library using a non-default Session
mySession = session("C:/Program Files/SyncroSim/1/SyncroSim.Console.exe")
myLib = ssimLibrary(name="st-sim",model="st-sim",cSession=mySession)
session(myLib)

showMethods(class="Session",where=loadNamespace("rsyncrosim")) #Methods for the Session
filepath(mySession)   # Lists the folder location of syncrosim session
version(mySession)   # Lists the version of syncrosim session
modules(mySession)   # Dataframe of the modules installed with this version of syncrosim, listing all of its properties as columns

#add and remove modules
TO DO:

#create or load a library using a default Session
myLib = ssimLibrary(name="st-sim",model="st-sim")
session(myLib)
}

