% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{session<-}
\alias{session<-}
\alias{session<-,character-method}
\alias{session<-,SsimObject-method}
\title{Set a SyncroSim Session.}
\usage{
session(ssimObject) <- value

\S4method{session}{character}(ssimObject) <- value

\S4method{session}{SsimObject}(ssimObject) <- value
}
\arguments{
\item{ssimObject}{SsimObject, \code{\link{Project}} or \code{\link{Scenario}}.}

\item{value}{A SyncroSim \code{\link{Session}}.}
}
\value{
Returns a SyncroSim object containing a Session.
}
\description{
Set the Session of a \code{\link{SsimLibrary}}, \code{\link{Project}} or
\code{\link{Scenario}} object.
}
\details{
In order to avoid problems with SyncroSim version compatibility and library
updating, the new session must have the same filepath as the session of the
SsimObject e.g. filepath(value)==filepath(session(ssimObject))
}
