% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{session<-}
\alias{session<-}
\title{Set a SyncroSim session.}
\usage{
session(ssimObject) <- value
}
\arguments{
\item{ssimObject}{SsimObject/Project/Scenario.}

\item{value}{A SyncroSim Session.}
}
\value{
An SyncroSim object containing a Session.
}
\description{
Set the Session of a SsimLibrary, Project or Scenario object.
}
\details{
In order to avoid problems with SyncroSim version compatibility and library updating, 
the new session must have the same filepath as the session of the SsimObject e.g. filepath(value)==filepath(session(ssimObject))
}
\examples{
myLibrary = ssimLibrary()
session(myLibrary)=session()
session(myLibrary)
}

