% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{session<-}
\alias{session<-}
\title{Set a SyncroSim session.}
\usage{
session(x) <- value
}
\arguments{
\item{x=}{A SyncroSim Session.}
}
\value{
An SyncroSim object containing a Session.
}
\description{
Set the Session of a SSimLibrary, Project or Scenario object.
}
\examples{
myLib = ssimLibrary()
session(ssimLibrary)<-session()
session(ssimLibrary)
}

