% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{session}
\alias{session}
\title{Start or get a SyncroSim session.}
\usage{
session(x = NULL, ...)
}
\arguments{
\item{x}{A path to SyncroSim.Console.exe or an object containing a Session.
If NULL the usual locations are searched.}

\item{silent}{Applies only if x is a path or NULL. If TRUE, warnings from the console are ignored. Otherwise they are printed.}
}
\value{
An SyncroSim Session object containing a valid console path.
}
\description{
Methods to create a Syncrosim session or fetch one from a SSimLibrary, Project or Scenario object.
}
\examples{
#Look for SyncroSim in the usual places
mySession = session()
path(mySession)

#Specify a SyncroSim version
mySession = session("C:/Program Files/SyncroSim/1/SyncroSim.Console.exe")

#Get the session from an SSimLibrary
myLib = ssimLibrary(name="st-sim",model="st-sim")
session(myLib)

#assign a session to a SyncroSim library
session(myLib)=session()
}

