% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/session.R, R/ssimLibrary.R
\docType{methods}
\name{session}
\alias{session}
\alias{session,SSimLibrary-method}
\alias{session,missingOrNULLOrChar-method}
\title{Start or get a SyncroSim session.}
\usage{
session(x = NULL)

\S4method{session}{missingOrNULLOrChar}(x = NULL)

\S4method{session}{SSimLibrary}(x = NULL)
}
\arguments{
\item{x=NULL}{A path to SyncroSim.Console.exe or an object containing a Session.
If NULL the usual locations are searched.}
}
\value{
An SyncroSim Session object containing a valid console path.
}
\description{
Methods to create a Syncrosim session or fetch one from a SSimLibrary, Project or Scenario object.
}
\section{Methods (by class)}{
\itemize{
\item \code{missingOrNULLOrChar}: Create a SyncroSim Session from a path or get default Session.

\item \code{SSimLibrary}: Session from an SSimLibrary.
}}
\examples{
#Look for SyncroSim in the usual places
mySession = session()
path(mySession)

#Specify a SyncroSim version
mySession = session("C:/Program Files/SyncroSim/1/SyncroSim.Console.exe")

#Get the session from an SSimLibrary
myLib = ssimLibrary(name="st-sim",model="st-sim")
session(myLib)
}

