% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{session}
\alias{session}
\alias{session,missingOrNULLOrChar-method}
\alias{session,SsimObject-method}
\title{Creates or returns a SyncroSim session.}
\usage{
session(x = NULL, silent = T, printCmd = F)

\S4method{session}{missingOrNULLOrChar}(x = NULL, silent = T, printCmd = F)

\S4method{session}{SsimObject}(x = NULL, silent = T, printCmd = F)
}
\arguments{
\item{x}{Character or SsimObject. An optional path to the SyncroSim installation.}

\item{silent}{Logical. Applies only if x is a path or NULL. If TRUE, warnings from the console are ignored. Otherwise they are printed.}

\item{printCmd}{Logical. Applies only if x is a path or NULL. If TRUE, arguments passed to the SyncroSim console are also printed. Helpful for debugging. FALSE by default.}
}
\value{
A SyncroSim Session object.
}
\description{
Methods to create or return a Syncrosim session.
}
\examples{
#Create a library using a default Session and base package
myLib = ssimLibrary(name="mylib")

#Create a library using a non-default Session
mySession = session("C:/Downloads/SyncroSim")
myLib = ssimLibrary(name="mylib",session=mySession)

filepath(mySession)     # Lists the folder location of syncrosim session
version(mySession)      # Lists the version of syncrosim session
package(mySession)      # Dataframe of the packages installed with this version of syncrosim.
basePackage(mySession)  # Dataframe of the base packages installed with this version of syncrosim.
}
