% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signout.R
\name{signout}
\alias{signout}
\title{Signs out of SyncroSim}
\usage{
signout(session = NULL)
}
\arguments{
\item{session}{\code{\link{Session}} object. If \code{NULL}(default), the
default session will be used}
}
\value{
Character string: whether sign out was successful or not.
}
\description{
Signs out of syncrosim.com.
}
\examples{
\dontrun{
# Sign out of SyncroSim session
mySession <- session()
signout(mySession)
}

}
