% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{silent<-}
\alias{silent<-}
\title{Set silent property of a Session}
\usage{
silent(session) <- value
}
\arguments{
\item{session}{Session}

\item{value}{logical}
}
\description{
Set silent property of a sessio to TRUE or FALSE
}

