% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silent.R
\name{silent<-}
\alias{silent<-}
\alias{silent<-,character-method}
\alias{silent<-,Session-method}
\title{Set silent property of a Session}
\usage{
silent(session) <- value

\S4method{silent}{character}(session) <- value

\S4method{silent}{Session}(session) <- value
}
\arguments{
\item{session}{Session.}

\item{value}{Logical.}
}
\value{
The updated ssimObject.
}
\description{
Set silent property of a \code{\link{Session}} to TRUE or FALSE.
}
