% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{silent}
\alias{silent}
\title{Check if a Session is silent}
\usage{
silent(x)
}
\arguments{
\item{x}{A SyncroSim \code{\link{Session}} object.}
}
\description{
Checks whether a SyncroSim Session is silent or not.
}

