% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silent.R
\name{silent}
\alias{silent}
\alias{silent,Session-method}
\alias{silent,missingOrNULLOrChar-method}
\title{Check if a Session is silent}
\usage{
silent(session)

\S4method{silent}{Session}(session)

\S4method{silent}{missingOrNULLOrChar}(session)
}
\arguments{
\item{session}{Session or character. An object of class \code{\link{Session}}
or path to a session. If NULL, the default session will be used.}
}
\value{
Returns a logical value: \code{TRUE} if the session is silent and \code{FALSE} otherwise.
}
\description{
Checks whether a SyncroSim \code{\link{Session}} is silent or not. In a silent
session, warnings from the console are ignored.
}
\examples{
\donttest{
temp_dir <- tempdir()
mySession <- session()

silent(mySession)
}

}
