% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{spatialData}
\alias{spatialData}
\title{Get spatial inputs or outputs from a SyncroSim scenario.}
\usage{
spatialData(x, sheet, iterations = NULL, timesteps = NULL)
}
\arguments{
\item{x}{A SyncroSim results Scenario or list of SyncroSim result Scenarios.}

\item{sheet}{The name of a spatial datasheet. See subset(datasheets(myResultScenario),isSpatial)$name for options.}

\item{iterations}{A vector of iterations. If NULL(default) all available iterations will be included}

\item{timesteps}{A vector of timesteps. If NULL(default) all available timesteps will be included.}
}
\value{
A RasterStack or RasterBrick object. See raster package documentation for details.
}
\description{
Get spatial inputs or outputs from a SyncroSim scenario.
}

