% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlStatements.R
\name{sqlStatements}
\alias{sqlStatements}
\title{Get SELECT and GROUP BY Statements}
\usage{
sqlStatements(varNames, drop = NULL, aggregate = NULL,
  aggregateFunction = "SUM")
}
\arguments{
\item{varNames}{Vector of all variables.}

\item{aggregateFunction}{An SQL aggregate function (e.g. SUM, COUNT)}

\item{groupVariables}{Vector of variables to GROUP BY.}

\item{aggregateVariables}{Vector of variables to aggregate using aggregateFunction}
}
\value{
A list of SQL SELECT and GROUP BY statements.
}
\description{
Creates SELECT and GROUP BY SQL Satements.
Variables are column names.
}

