% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{envOutputFolder}
\alias{envOutputFolder}
\title{SyncroSim DataSheet Output Folder}
\usage{
envOutputFolder(scenario, datasheetName)
}
\arguments{
\item{scenario}{Scenario.  A SyncroSim result scenario.}

\item{datasheetName}{character.  The output datasheet name.}
}
\value{
a folder name for the specified data sheet
}
\description{
Creates and returns a SyncroSim DataSheet Output Folder
}
