% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{envReportProgress}
\alias{envReportProgress}
\alias{envBeginSimulation}
\alias{envStepSimulation}
\alias{envEndSimulation}
\title{Reports progress for a SyncroSim simulation}
\usage{
envReportProgress(iteration, timestep)

envBeginSimulation(totalSteps)

envStepSimulation()

envEndSimulation()
}
\arguments{
\item{iteration}{integer.  The current iteration.}

\item{timestep}{integer.  The current timestep.

#' @return
No returned value, used for side effects.}

\item{totalSteps}{integer.  The total number of steps in the simulation.

#' @return
No returned value, used for side effects.}
}
\value{
No returned value, used for side effects.

\donttest{
envStepSimulation()
}

No returned value, used for side effects.
}
\description{
This function is part of a set of functions designed to facilitate the
development of R-based Syncrosim Packages. This function reports progress 
for a SyncroSim simulation.

This function is part of a set of functions designed to facilitate the
development of R-based SyncroSim Packages. Specifically, this function begins
the reporting of a SyncroSim simulation.

This function is part of a set of functions designed to facilitate the
development of R-based Syncrosim Packages. This function steps a SyncroSim 
simulation, and is to be called between \link{\code{envBeginSimulation}} and 
\link{\code{envEndSimulation}}.

This function is part of a set of functions designed to facilitate the
development of R-based Syncrosim Packages. This function ends a SyncroSim 
simulation.
}
\examples{
\donttest{
envReportProgress()
}

\donttest{
n_iter <- 50
n_timesteps <- 10
n_steps <- n_iter * n_timesteps
envBeginSimulation(n_steps)
}

\donttest{
envEndSimulation()
}

}
