% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{ssimEnvironment}
\alias{ssimEnvironment}
\title{SyncroSim Environment}
\usage{
ssimEnvironment()
}
\value{
Returns a single-row data.frame of SyncroSim specific environment variables.
}
\description{
This function is part of a set of functions designed to facilitate the
development of R-based Syncrosim Packages. \code{ssimEnvironment} retrieves
specific environment variables.
}
\examples{
\donttest{
# Get the whole set of variables
e <- ssimEnvironment()

# Get the path to transfer direcyory, for instance
transferdir <- e$TransferDirectory
}

}
