% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{ssimInputFolder}
\alias{ssimInputFolder}
\title{SyncroSim DataSheet Input Folder}
\usage{
ssimInputFolder(scenario, datasheetName)
}
\value{
a folder name for the specified data sheet
}
\description{
Retrieves a SyncroSim DataSheet Input Folder
}
