% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimLevelplot.R
\docType{methods}
\name{ssimLevelplot}
\alias{ssimLevelplot}
\alias{ssimLevelplot,RasterLayer-method}
\title{levelplot of a categorical RasterLayer with a Raster Attribute Table}
\usage{
ssimLevelplot(raster, attribute, ...)

\S4method{ssimLevelplot}{RasterLayer}(raster, attribute, ...)
}
\arguments{
\item{raster}{A RasterLayer with a raster attribute table set by spatialData() or ssimRatify().}

\item{attribute}{character string. The attribute to be plotted. This must be a column name in the raster attribute table.}
}
\description{
Plot a RasterLayer with raster attributes set by spatialData() or ssimRatify().
This is a wrapper around the levelplot() function of the rasterVis package.
}
\section{Methods (by class)}{
\itemize{
\item \code{RasterLayer}: Make a categorical levelplot of a RasterLayer.
}}
\examples{

levels(myRaster) #Retrieve the raster attribute table
ssimLevelplot(myRaster,attribute="Name") #plot Name attribute

}

