% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAClassDefinitions.R
\docType{class}
\name{SsimLibrary-class}
\alias{SsimLibrary-class}
\title{SyncroSim Library class}
\description{
\code{SsimLibrary} object representing a SyncroSim Library. A SsimLibrary is the
highest level of organization in the SyncroSim workflow and contains at
least one \code{\link{Project}}.
}
\section{Slots}{

\describe{
\item{\code{session}}{\code{\link{Session}} object}

\item{\code{filepath}}{character string. The path to the SsimLibrary on disk}

\item{\code{datasheetNames}}{character string. The name and scope of all Datasheets
in the SsimLibrary.}
}}

\seealso{
See \code{\link{ssimLibrary}} for options when creating or loading a
SyncroSim SsimLibrary.
}
