% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAClassDefinitions.R
\docType{class}
\name{SsimLibrary-class}
\alias{SsimLibrary-class}
\alias{SsimLibrary}
\title{SyncroSim Library class}
\description{
\code{SsimLibrary} object representing a SyncroSim Library.
}
\section{Slots}{

\describe{
\item{\code{session}}{The SyncroSim Session.}

\item{\code{filepath}}{The path to the Library on disk.}

\item{\code{datasheetNames}}{The name and scope of all datasheets in the Library.}
}}

\seealso{
See \code{\link{ssimLibrary}} for options when creating or loading a SyncroSim Library.
}
