% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAClassDefinitions.R
\docType{class}
\name{SsimLibrary-class}
\alias{SsimLibrary}
\alias{SsimLibrary-class}
\title{SyncroSim Library class}
\description{
\code{SsimLibrary} object representing a SyncroSim Library.
}
\section{Slots}{

\describe{
\item{\code{session}}{The SyncroSim Session.}

\item{\code{filepath}}{The path to the Library on disk.}

\item{\code{datasheetNames}}{The name and scope of all datasheets in the Library.}
}}
<<<<<<< HEAD
\examples{
#TODO - update examples
# Create or load and query a SyncroSim Library.
myLibrary = ssimLibrary()
session(myLibrary)
filepath(myLibrary)
info(myLibrary)

# Add or load a project, then get the SyncroSim Library associated with that Project
myProject = project(myLibrary,project="a project")

myLibrary = ssimLibrary(myProject)

}
||||||| merged common ancestors

\examples{
#TODO - update examples
# Create or load and query a SyncroSim Library.
myLibrary = ssimLibrary()
session(myLibrary)
filepath(myLibrary)
info(myLibrary)

# Add or load a project, then get the SyncroSim Library associated with that Project
myProject = project(myLibrary,project="a project")

myLibrary = ssimLibrary(myProject)

}
=======

>>>>>>> ca16fc4b3ec3b292b67f4eda8481627a2c390a4f
\seealso{
See \code{\link{ssimLibrary}} for options when creating or loading a SyncroSim Library.
}

