% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAClassDefinitions.R
\docType{class}
\name{SsimLibrary-class}
\alias{SsimLibrary}
\alias{SsimLibrary-class}
\title{SyncroSim Library class}
\description{
\code{SsimLibrary} object representing a SyncroSim Library.
}
\section{Slots}{

\describe{
\item{\code{session}}{The SyncroSim Session.}

\item{\code{filepath}}{The path to the library on disk.}

\item{\code{datasheetNames}}{The names and scope of all datasheets in the library. Used to speed calculations.}
}}
\examples{
#TODO - update examples
# Create or load and query a SyncroSim Library.
myLibrary = ssimLibrary()
session(myLibrary)
filepath(myLibrary)
info(myLibrary)

# Add or load a project, then get the SyncroSim Library associated with that Project
myProject = project(myLibrary,project="a project")

myLibrary = ssimLibrary(myProject)

}
\seealso{
See \code{\link{ssimLibrary}} for options when creating or loading an SyncroSim library.
}

