% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimLibrary.R
\docType{class}
\name{SSimLibrary-class}
\alias{SSimLibrary}
\alias{SSimLibrary-class}
\title{SyncroSim Library class}
\description{
\code{SSimLibrary} object representing a SyncroSim Library.
}
\section{Slots}{

\describe{
\item{\code{session}}{The SyncroSim session.}

\item{\code{filepath}}{The path to the library on disk.}

\item{\code{datasheetNames}}{The names and scope of all datasheets in the library. Used to speed calculations.}
}}
\examples{
# Create or load and query a SyncroSim Library.
myLibrary = ssimLibrary(model="stsim")
session(myLibrary)
filepath(myLibrary)
info(myLibrary)

# Add or load a project, then get the SyncroSim Library associated with that Project
myProject = project(myLibrary)
myLibrary = ssimLibrary(myProject)

}
\seealso{
See \code{\link{ssimLibrary}} for options when creating or loading an SyncroSim library.
}

