% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimLibrary.R
\docType{class}
\name{SSimLibrary-class}
\alias{SSimLibrary}
\alias{SSimLibrary-class}
\title{SyncroSim Library class}
\description{
\code{SSimLibrary} object representing a SyncroSim Library.
}
\section{Slots}{

\describe{
\item{\code{session}}{The SyncroSim session.}

\item{\code{filepath}}{The path to the library on disk.}

\item{\code{info}}{Basic library properties.}
}}
\examples{
#Create or load and query an STSim library.
myLib = ssimLibrary(model="st-sim")
session(myLib)
filepath(myLib)
info(myLib)
}
\seealso{
See \code{\link{ssimLibrary}} for options when creating or loading an SyncroSim library.
}

