% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/ssimLibrary.R, R/project.R, R/scenario.R
\docType{methods}
\name{ssimLibrary}
\alias{ssimLibrary}
\alias{ssimLibrary,Project-method}
\alias{ssimLibrary,Scenario-method}
\title{Create or open a library.}
\usage{
ssimLibrary(model = NULL, ...)

\S4method{ssimLibrary}{missingOrNULLOrChar}(model = NULL, name = NULL,
  session = NULL, addons = NULL, backup = F, backupName = "backup",
  backupOverwrite = T, forceUpdate = F)

\S4method{ssimLibrary}{Project}(model)

\S4method{ssimLibrary}{Scenario}(model)
}
\arguments{
\item{model}{A model type or a SyncroSim Project or Scenario. Optional when loading an existing library using a name.}

\item{name}{A library file name or library file path. If not a path library is created or opened in the current working directory.}

\item{session}{A SyncroSim \code{Session}. If NULL, the default SyncroSim Session will be used.}

\item{addons}{One or more addons. See addons() for options.}

\item{backup}{If TRUE, a backup copy is made when an existing library is opened.}

\item{backupName}{Added to a library filepath to create a backup library.}

\item{backupOverwrite}{If TRUE, the existing backup of a library (if any) will be overwritten.}

\item{forceUpdate}{If FALSE (default) user will be prompted to approve any required updates. If TRUE, required updates will be applied silently.}
}
\value{
An \code{SSimLibrary} object representing a SyncroSim library.
}
\description{
Creates or opens an \code{\link{SSimLibrary}} object representing a SyncroSim library.
}
\details{
\itemize{
  \item {If model is SyncroSim Project or Scenario: }{Returns the \code{\link{SSimLibrary}} associated with the Project or Scenario.}
  \item {If given no name and no model: }{Opens an existing SyncroSim library in
  the current working directory - returns an error if more than one library exists. If library does not exist and only one model is installed - creates a library of that type.}
  \item {If given a model but no name: }{Opens or creates a library called <model>.ssim in the current working directory.}
  \item {If given a name but no model: }{Attempts to open a library of that name. Returns an error if that library does not already exist.}
  \item {If given a name and a model: }{Opens or creates a library called <name>.ssim. Returns an error if the library already exists but is a different type of model.}
}
}
\section{Methods (by class)}{
\itemize{
\item \code{Project}: Get the SSimLibrary associated with a SyncroSim Project.

\item \code{Scenario}: Get the SSimLibrary associated with a SyncroSim Scenario.
}}
\examples{
# See the installed models
models(session())

# Create a library called <model>.ssim in the current working directory.
myLibrary = ssimLibrary(model="stsim")
session(myLibrary) #The SycroSim session
filepath(myLibrary) #Path to the file on disk.
info(myLibrary) #Model type and other library information.

# Open an existing SyncroSim library in the current working directory - don't make a backup copy.
myLibrary = ssimLibrary()

# Create a library with a name in the current working directory
mySecondLibrary = ssimLibrary(name="Lib2",model="stsim")

# Create a library with a name in another directory
myThirdLibrary = ssimLibrary(name=paste0(getwd(),"/Temp/Lib3"),model="stsim")

# Create or load a library using a specific session
mySession = session("C:/Program Files/SyncroSim/1/SyncroSim.Console.exe")
myLibrary = ssimLibrary(name="Lib2",session=mySession)

# Add a project and get the library associated with that project
myProject = project(myLibrary)
myLibrary = ssimLibrary(myProject)
}

