% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimLibrary.R
\name{ssimLibrary}
\alias{ssimLibrary}
\title{Create or open a library.}
\usage{
ssimLibrary(model = NULL, name = NULL, cSession = NULL, ...)
}
\arguments{
\item{model}{The model type. Optional when loading an existing library.}

\item{name}{A library file name or library file path. If not a path library is created or opened in the current working directory.}

\item{cSession}{A SyncroSim \code{Session}. If NULL, the default SyncroSim Session will be used.}
}
\value{
An \code{SSimLibrary} object representing a SyncroSim library.
}
\description{
Creates or opens an \code{\link{SSimLibrary}} object representing a SyncroSim library.
}
\details{
\itemize{
  \item {If given no name and no model: }{Opens an existing SyncroSim library in
  the current working directory - returns an error if more than one library exists.}
  \item {If given a model but no name: }{Opens or creates a library called <model>.ssim in the current working directory.}
  \item {If given a name but no model: }{Attempts to open a library of that name. Returns an error if that library does not already exist.}
  \item {If given a name and a model: }{Opens or creates a library called <name>.ssim. Returns an error if the library already exists but is a different type of model.}
}
}
\examples{
#Create a library called <model>.ssim in the current working directory.
myLib = ssimLibrary(model="st-sim")
session(myLib) #The SycroSim session
filepath(myLib) #Path to the file on disk.
info(myLib) #Model type and other library information.

#Open an existing SyncroSim library in the current working directory.
myLib = ssimLibrary()

#Create a library with name in the current working directory
myLib2 = ssimLibrary(name="Lib2",model="st-sim")

#Create a library with a name and model in another directory
myLib3 = ssimLibrary(name=paste0(getwd(),"/Temp/Lib3"),model="st-sim")

#create or load a library using a specific session
mySession = session("C:/Program Files/SyncroSim/1/SyncroSim.Console.exe")
myLib = ssimLibrary(name="Lib2",cSession=mySession)
}

