% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{ssimOutputFolder}
\alias{ssimOutputFolder}
\title{SyncroSim DataSheet Output Folder}
\usage{
ssimOutputFolder(scenario, datasheetName)
}
\value{
a folder name for the specified data sheet
}
\description{
Retrieves a SyncroSim DataSheet Output Folder
}
