% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimUpdate.R
\name{ssimUpdate}
\alias{ssimUpdate}
\alias{ssimUpdate,character-method}
\alias{ssimUpdate,SsimObject-method}
\title{Apply updates}
\usage{
ssimUpdate(ssimObject)

\S4method{ssimUpdate}{character}(ssimObject)

\S4method{ssimUpdate}{SsimObject}(ssimObject)
}
\arguments{
\item{ssimObject}{An object of class \code{\link{Session}}, \code{\link{Project}}, 
or \code{\link{SsimLibrary}}.}
}
\value{
This function invisibly returns `TRUE` upon success (i.e.successful 
update) and `FALSE` upon failure.
}
\description{
Apply updates to a \code{\link{SsimLibrary}}, or a \code{\link{Project}} or 
\code{\link{Scenario}} associated with a Library.
}
\examples{
\donttest{
temp_dir <- tempdir()
mySession <- session()
myLibrary <- ssimLibrary(name = file.path(temp_dir,"testlib"), session = mySession)

# Update Library
ssimUpdate(myLibrary)

myProject <- project(myLibrary, name = "My Project")

# Update Project
ssimUpdate(myProject)

myScenario <- scenario(myLibrary, name = "MyScenario")

# Update scenario
ssimUpdate(myScenario)
}

}
