% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimUpdate.R
\docType{methods}
\name{ssimUpdate}
\alias{ssimUpdate}
\alias{ssimUpdate,SsimObject-method}
\alias{ssimUpdate,character-method}
\title{Apply updates.}
\usage{
ssimUpdate(ssimObject)

\S4method{ssimUpdate}{character}(ssimObject)

\S4method{ssimUpdate}{SsimObject}(ssimObject)
}
\arguments{
\item{ssimObject}{SsimLibrary/Project/Scenario}
}
\value{
"saved" or a failure message from the console.
}
\description{
Apply updates to a SyncroSim Library,or a Project or Scenario associated with a Library.
}

