% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimUpdate.R
\name{ssimUpdate}
\alias{ssimUpdate}
\title{Apply updates.}
\usage{
ssimUpdate(ssimObject)
}
\arguments{
\item{ssimObject}{SsimLibrary/Project/Scenario}
}
\value{
"saved" or a failure message from the console.
}
\description{
Apply updates to a SyncroSim Library,or a Project or Scenario associated with a Library.
}

