% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath.R
\docType{methods}
\name{tempfilepath}
\alias{tempfilepath}
\alias{tempfilepath,character-method}
\alias{tempfilepath,Session-method}
\alias{tempfilepath,SsimObject-method}
\title{The temporary file path to a SyncroSim object on disk}
\usage{
tempfilepath(ssimObject)

\S4method{tempfilepath}{character}(ssimObject)

\S4method{tempfilepath}{Session}(ssimObject)

\S4method{tempfilepath}{SsimObject}(ssimObject)
}
\arguments{
\item{ssimObject}{An object containing a filepath.}
}
\description{
The temporary file path to a SyncroSim Session, SSimLibarary, Project or Scenario on disk.
}
