% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstallPackage.R
\name{uninstallPackage}
\alias{uninstallPackage}
\alias{uninstallPackage,ANY,character-method}
\alias{uninstallPackage,ANY,missingOrNULL-method}
\alias{uninstallPackage,ANY,Session-method}
\title{Removes a package from SyncroSim installation}
\usage{
uninstallPackage(name, session = NULL, force = FALSE)

\S4method{uninstallPackage}{ANY,character}(name, session = NULL, force = FALSE)

\S4method{uninstallPackage}{ANY,missingOrNULL}(name, session = NULL, force = FALSE)

\S4method{uninstallPackage}{ANY,Session}(name, session = NULL, force = FALSE)
}
\arguments{
\item{name}{character. The name of the package to uninstall}

\item{session}{\code{\link{Session}} object. If \code{NULL} (default),
\code{session()} will be used}

\item{force}{logical. If \code{TRUE}, uninstall without requiring
confirmation from the user. Default is \code{FALSE}}
}
\value{
Invisibly returns \code{TRUE} upon success (i.e.successful
removal) and \code{FALSE} upon failure.
}
\description{
Removes a package from SyncroSim installation
}
\examples{
\donttest{
# Set SyncroSim Session
mySession <- session()

# Uninstalls package from SyncroSim Session
uninstallPackage("stsim", mySession, force = FALSE)
}

}
