% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimLibrary.R
\name{update}
\alias{update}
\title{Apply updates.}
\usage{
update(x)
}
\arguments{
\item{x}{An SSimLibrary object, or a Project or Scenario associated with a Library}
}
\value{
Success or a failure message from the console.
}
\description{
Apply updates to a SyncroSim Library.
}

