% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePackage.R
\name{updatePackage}
\alias{updatePackage}
\alias{updatePackage,ANY,character-method}
\alias{updatePackage,ANY,missingOrNULL-method}
\alias{updatePackage,ANY,Session-method}
\title{Update Package}
\usage{
updatePackage(name = NULL, session = NULL, listonly = FALSE)

\S4method{updatePackage}{ANY,character}(name = NULL, session = NULL, listonly = FALSE)

\S4method{updatePackage}{ANY,missingOrNULL}(name = NULL, session = NULL, listonly = FALSE)

\S4method{updatePackage}{ANY,Session}(name = NULL, session = NULL, listonly = FALSE)
}
\arguments{
\item{name}{Character string.  The name of the package to update. 
If NULL, all packages will be updated.}

\item{session}{An object of class \code{\link{Session}}.}

\item{listonly}{Logical. If TRUE, available updates are listed only.}
}
\value{
This function invisibly returns `TRUE` upon success (i.e.successful update)
and `FALSE` upon failure.
}
\description{
Updates a SyncroSim package.
}
\examples{
\donttest{
temp_dir <- tempdir()
mySession <- session()

updatePackage(name = "stsim", session = mySession, listonly = FALSE)
updatePackage(name = "stsim", session = mySession, listonly = TRUE)
}

}
