% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimEnvironment.R
\name{updateRunLog}
\alias{updateRunLog}
\title{Function to write to the SyncroSim run log}
\usage{
updateRunLog(..., sep = "", type = "status")
}
\arguments{
\item{...}{One or more objects which can be coerced to character
which are pasted together using \code{sep}.}

\item{sep}{character. Used to separate terms. Not NA_character_}

\item{type}{character. Type of message to add to run log. One of "status",
"info", or "warning".}
}
\value{
No returned value, used for side effects
}
\description{
This function is designed to facilitate the development of R-based Syncrosim
Packages by allowing developers to send messages to the run log.
}
\examples{
\dontrun{
# Write a message to run log
updateRunLog(msg)

# Construct and write a message to run log
updateRunLog(msg, additionalMsg, sep = " ")
}

}
