% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useConda.R
\name{useConda}
\alias{useConda}
\alias{useConda,character-method}
\alias{useConda,SsimLibrary-method}
\alias{useConda<-}
\alias{useConda<-,character-method}
\alias{useConda<-,SsimLibrary-method}
\title{Conda configuration of a SsimLibrary}
\usage{
useConda(ssimObject)

\S4method{useConda}{character}(ssimObject)

\S4method{useConda}{SsimLibrary}(ssimObject)

useConda(ssimObject) <- value

\S4method{useConda}{character}(ssimObject) <- value

\S4method{useConda}{SsimLibrary}(ssimObject) <- value
}
\arguments{
\item{value}{logical, string, or list of strings for whether to use Conda 
environments for SyncroSim packages used by the Library. If set to 
\code{TRUE}, then Conda environments will be used for all SyncroSim packages.}

\item{ssimLibrary}{\code{\link{SsimLibrary}} object}
}
\value{
A logical: the Conda configuration of the SsimLibrary.
}
\description{
Retrieves or sets the Conda configuration of a \code{\link{SsimLibrary}}.
}
\examples{
\donttest{
# Specify file path and name of new SsimLibrary
myLibraryName <- file.path(tempdir(), "testlib")

# Set up a SyncroSim Session, SsimLibrary, Project, and Scenario
mySession <- session()
myLibrary <- ssimLibrary(name = myLibraryName, session = mySession)

# Retrieve Conda configuration status of the SsimLibrary
useConda(myLibrary)

# Set the Conda configuration of the SyncroSim Library
useConda(myLibrary) <- TRUE

# Only use Conda with the specified SyncroSim packages
useConda(myLibrary) <- "helloworld"

# Only use Conda with multiple specified SyncroSim packages
useConda(myLibrary) <- ["helloworld", "stsim"]
}

}
