% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{version}
\alias{version}
\title{The SyncroSim version}
\usage{
version(x)
}
\arguments{
\item{x}{A SyncroSim \code{\link{Session}} object.}
}
\description{
The version of a SyncroSim \code{\link{Session}}.
}

