% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{version}
\alias{version}
\alias{version,character-method}
\alias{version,missingOrNULL-method}
\alias{version,Session-method}
\title{Retrieves SyncroSim version}
\usage{
version(session = NULL)

\S4method{version}{character}(session = NULL)

\S4method{version}{missingOrNULL}(session = NULL)

\S4method{version}{Session}(session = NULL)
}
\arguments{
\item{session}{\code{\link{Session}} object}
}
\value{
A character string e.g. "2.2.13".
}
\description{
Retrieves the version of a SyncroSim Session.
}
\examples{
\donttest{
# Set SyncroSim Session
mySession <- session()

# Retrieve version of SyncroSim associated with Session
version(mySession)
}

}
