% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{version}
\alias{version}
\alias{version,character-method}
\alias{version,missingOrNULL-method}
\alias{version,Session-method}
\title{The SyncroSim version}
\usage{
version(session = NULL)

\S4method{version}{character}(session = NULL)

\S4method{version}{missingOrNULL}(session = NULL)

\S4method{version}{Session}(session = NULL)
}
\arguments{
\item{session}{Session.}
}
\description{
The version of a SyncroSim Session.
}
