% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\docType{methods}
\name{version}
\alias{version}
\alias{version,Session-method}
\alias{version,missingOrNULL-method}
\title{The SyncroSim version}
\usage{
version(session = NULL)

\S4method{version}{missingOrNULL}(session = NULL)

\S4method{version}{Session}(session = NULL)
}
\arguments{
\item{session}{Session.}
}
\description{
The version of a SyncroSim Session
}
\section{Methods (by class)}{
\itemize{
\item \code{missingOrNULL}: Version of the default Session.

\item \code{Session}: Version of a Session.
}}

