% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewProfile.R
\name{viewProfile}
\alias{viewProfile}
\title{Display SyncroSim profile}
\usage{
viewProfile(session = NULL, ...)
}
\arguments{
\item{session}{\code{\link{Session}} object. If \code{NULL}(default), the
default session will be used}
}
\description{
Displays the currently signed in SyncroSim profile information. To sign in
to SyncroSim use the \code{\link{signin}} function.
}
\examples{
\dontrun{
# Retrieve profile information for a SyncroSim session
mySession <- session()
viewProfile(mySession)
}

}
